Usage

Climwip example from Brunner et al. (2020)

Weights for models of lgm experiment

Load model data for experiment lgm and historical in one run from new config file

TODO: show that we get the same outfit different approaches!

Filtering data

General options

Additional options for data preprocessed with ESMValTool

Getting started

How to load data

How to compute weights

Usage

Data Representation

Weights

Application of weights

Functions

Index

Loading data

Loading data from directories

Loading data preprocessed with ESMValTool

References

Requirements

Computing weights

Area-weighted RMSE

Generalized distances

Computation of overall weight for model $i$

Computation of performance/independence weights only

The purpose of this package is twofold: on the one hand, we built it to handle preprocessed Earth System data easily within Julia and, on the other hand, to compute weights for different sets of Earth System Models following the approach from Brunner et al (2020) and adding the possibility to use and combine different performance metrics.

Data Representation

The basic representation of data is in form of a dictionary that simply maps from ids (Strings) to the data itself stored as YAXArrays. We name this type alias a DataMap.

The metadata of each dataset ($datamap[id].properties$) contains the metadata loaded from the original files as well as additional information if it is provided in the optional argument meta_data.

Weights

<!– $julia import ModelWeights as mw weights = mw.computeWeights(model_data_historical, obs_data, config_weights);$

The Weights object has several fields that each store a DimensionalData.DimArray. Some of them refer to normalized weights, others refer to the distances on the basis of which the weights were computed.

performance_distances and independence_distances contain the distances

between models and data (performance) or between models and models (independence) for every combination of variable and diagnostic/statistic.

Di and Sij are the generalized distances, i.e. weighted average of all

distances across variables and diagnostic. So, $D_i$ is a vector with size 1xn where n is the number of models (not on level of members!) and $S_{ij}$ is a matrix of size n x n.

wP and wI respectively store the normalized performance/independence weights,

both have size 1xn (n: number of models).

w is the overall weight vector of length n. –>

Application of weights

Add example of how weights are applied to data

Functions

ModelWeights.Data.addDiagnostic! — Method

addDiagnostic!(datamap::DataMap, fn::Function, args...; kwargs..., id_suffix::String="")

Compute a diagnostic by running fn with positional arguments args and keyword arguments kwargs and add result to datamap at the id of the computed result concatenated with #id_suffix if id_suffix is not empty.

ModelWeights.Data.addDiagnostic! — Method

addDiagnostic!(
    data::ClimateData, fn::Function, ids::Vector{String}, key_suffix::String, args...; kwargs...
)

Run addDiagnostic! for model and observational data in data.

ModelWeights.Data.alignPhysics — Method

alignPhysics(
    datamap::DataMap, members::Vector{String}; level_shared::Union{Level, Nothing}=nothing
)

Return new DataMap with only the models retained that share the same physics as the respective model's members in members.

If level_shared is set, resulting DataMap is subset accordingly.

ModelWeights.Data.anomalies — Method

anomalies(orig_data::YAXArray, ref_data::YAXArray; stats_name::String="ANOM")

Compute difference of orig_data and ref_data.

The id of the original data and of the reference data is added to the metadata.

ModelWeights.Data.approxAreaWeights — Method

approxAreaWeights(latitudes::Vector{<:Number})

Create a YAXArray with the cosine of latitudes which approximates the cell area on the respective latitude.

ModelWeights.Data.areaWeightedRMSE — Method

areaWeightedRMSE(m1::AbstractArray, m2::AbstractArray, mask::AbstractArray)

Compute the area weighted (approximated by cosine of latitudes in radians) root mean squared error between m1 and m2.

Arguments:

m1: must have dimensions 'lon', 'lat'.

m2: must have dimensions 'lon', 'lat'.

mask: has values 0,1. Locations where mask is 1 get a weight of 0.

ModelWeights.Data.checkConstraint — Method

checkConstraint(constraint::Union{Dict, Nothing})

Throw error if entries (except for level_shared) do not map to a vector of Strings.

Used to provide specific error messages of values that were not specified correctly.

ModelWeights.Data.combineAll — Method

combineAll(v::Vararg{Vector{String}}; sep::String="_")

Generate vector with all possible combinations of strings, each of which is a concatenation of elements from each input vector in the given order, concatenated by sep (default: _).

Arguments

v::Vararg{Vector{String}}: A variable number of input vectors.

sep::String: seperator

Example

julia> ModelWeights.Data.combineAll(["tos", "tas"], ["CLIM"])
2-element Vector{String}:
 "tos_CLIM"
 "tas_CLIM"

ModelWeights.Data.combineModelsFromMultipleFiles — Method

combineModelsFromMultipleFiles(
    data::Vector{<:YAXArray}; 
    names::Vector{String} = Vector{String}(),
    meta::Union{Dict{String, T}, Nothing} = nothing,
    dtype::String = "undef",
    new_dim::Symbol = :model,
    sorted::Bool = true

) where T <: Any

Combine data from different files into a single YAXArray. The meta data of the returned YAXArray is a combination of all values from all datasets using vectors with missing values for datasets that didn't have the respective metadata entry in their metadata, further combined with the properties in meta if provided.

All elements in data must share all dimensions except for time (if present). For timeseries data, the time dimension may cover different ranges. In that case the maximal timeseries is used and filled with NaN for missing values.

The combined YAXArray has the additional dimension new_dim (default: :model) with names as values. If names is not provided, default values 'model1', 'model2, etc. are used instead.

If sorted is true, model dimension of returned data is sorted alphabetically and the vector entries in the metadata dictionary of the returned array are also sorted accordingly.

ModelWeights.Data.constrainMetaData! — Method

constrainMetaData!(meta_attributes::Vector{MetaData}, constraint::Dict)

Subset entries in meta_attributes so that only those with properties specified in constraint remain.

Arguments

constraint::Dict: Mapping to vector specifiying the properties of which at least one

must be present for an id to be retained.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    path_data_dir::String, 
    id::String;
    meta_data::Dict{String, T} = Dict{String, Any}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "undef",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entry id with the data (model and/or obs depending on dtype) from all .nc files in paths and all .nc files in all directories in paths, possibly constraint by constraint.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    paths_data_dirs::Vector{String}, 
    data_ids::Vector{String};
    meta_data::Vector{Dict{String, T}} = Vector{Dict{String, Any}}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "undef",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entries data_ids with the data (model and/or obs depending on dtype) from all .nc files in all directories in paths_dirs, possibly constraint by constraint.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    paths_data_dirs::Vector{String}, 
    data_ids::Vector{String};
    meta_data::Vector{Dict{String, T}} = Vector{Dict{String, Any}}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "undef",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entries data_ids with the data (model and/or obs depending on dtype) from all .nc files in all directories in paths_dirs, possibly constraint by constraint.

For every loaded dataset (entry in built DataMap), files are loaded from several directories; each entry in paths_data_dirs points the the vector of data directories from where data is loaded for that dataset.

ModelWeights.Data.distancesData — Method

distancesData(model_data::DataMap, obs_data::DataMap, diagnostics_ids::Vector{String})

Compute RMSEs between models and observations for diagnostics.

Arguments:

diagnostics_ids::Vector{String}:

ModelWeights.Data.distancesData — Method

distancesData(models::YAXArray, observations::YAXArray)

Compute the distance as the area-weighted RMSE between model predictions and observations.

If several observational datasets are present, the average across all is taken.

ModelWeights.Data.distancesModels — Method

distancesModels(modelData::YAXArray)

Compute the area weighted RMSE between model predictions for each pair of models.

Arguments:

modelData::YAXArray: must have dimensions 'lon', 'lat', 'model' or 'member'.

ModelWeights.Data.filterPathsSharedModels — Method

filterPathsSharedModels(
    paths::Vector{String}, 
    shared_models::Vector{String}, 
    fn_format::Union{Symbol, String}
)

Every vector of paths in all_paths is filtered s.t. it only contains models or model members given in shared_models.

Arguments:

paths: contains paths to data files

ModelWeights.Data.filterPathsSharedModels — Method

filterPathsSharedModels(
    all_paths::Vector{Vector{String}}, 
    level_shared::Level, 
    fn_format::Union{Symbol, String}
)

Arguments:

all_paths: every entry refers to the paths to data files for the respective dataset

ModelWeights.Data.generalizedDistances — Method

generalizedDistances(distances_all::YAXArray, weights::DimArray)

For every variable in distances_all, compute the weighted sum of all diagnostics.

ModelWeights.Data.getAtModel — Method

getAtModel(data::YAXArray, dimension::Symbol, model::String)

Return data where dimension (member or model) has value model.

ModelWeights.Data.getByIdxModel — Method

getByIdxModel(data::YAXArray, dimension::Symbol, indices::Vector)

Return data where dimension (member or model) has value model.

ModelWeights.Data.getCMIPModelsKey — Method

getCMIPModelsKey(meta::Dict)

Return the respective key to retrieve model names in CMIP6 ('sourceid') and CMIP5 ('modelid') data.

If both keys are present, 'source_id' used in CMIP6 models is returned, if none is present, throw ArgumentError.

ModelWeights.Data.globalMeans — Method

globalMeans(data::YAXArray)

Return a YAXArray with area-weighted global means for each model in data.

Missing data is accounted for in the area-weights.

Arguments:

data::YAXArray: must have dimensions 'lon' and 'lat'.

ModelWeights.Data.individuatePath — Method

individuatePath(target::String)

If file at target path exists, concatenate target with current timestep, otherwise simply return target.

ModelWeights.Data.interpolatedWeightedQuantiles — Method

interpolatedWeightedQuantiles(
    quantiles::Vector{<:Number}, vals::Vector; weights=nothing
)

This implementation follows the one used by Brunner et al.

ModelWeights.Data.joinDicts — Method

joinDicts(v::Dict...; warn_msg::String="")

ModelWeights.Data.kelvinToCelsius! — Method

kelvinToCelsius!(datamap::DataMap)

Modify entries of datamap s.t. all data is given in Degree Celsius (instead) of Kelvin.

ModelWeights.Data.kelvinToCelsius — Method

kelvinToCelsius(data::AbstractArray)

Return a copy of data with values given in Kelvin covnerted into Degree Celsius.

ModelWeights.Data.loadDataFromESMValToolRecipes — Method

loadDataFromESMValToolRecipes(
    path_data::String,
    path_recipes::String;
    dir_per_var::Bool = true,
    constraint::Union{Dict, Nothing} = nothing,
    preview::Bool = false,
    sorted::Bool = true
)

Load the data from the ESMValTool recipes at path_recipes or, if preview is true, load meta data only.

Arguments:

path_data: top level directory were data is stored.

path_recipes: directory were ESMValTool recipes are stored; these must be the versions

in the run folder generated by ESMValTool named RECIPENAME_filled.yml.

dir_per_var: set to true (default) if there is a subdirectory in path_data for every

climate variable to be loaded.

constraint: TODO!

preview: set to true if only meta data should be loaded (default: false).

sorted: if true (default), the data is sorted alphabetically wrt model names.

ModelWeights.Data.loadDataFromYAML — Method

loadDataFromYAML(
    content::Dict;
    constraint::Union{Dict, Nothing} = nothing,
    preview::Bool = false,
    sorted::Bool = true,
    dtype::String = "undef"
)

Load a ClimateData-instance that contains the data specified in content, potentially constraint by values in constraint.

Arguments:

preview::Bool: if true (default: false), return metadata and corresponding paths without

actually loading any data.

ModelWeights.Data.loadDataMapCore — Method

loadDataMapCore(
    all_paths::Vector{Vector{String}},
    ids::Vector{String};
    constraint::Union{Dict, Nothing} = nothing,
    dtype::String = "undef",
    filename_format::Union{Symbol, String} = :cmip,
    sorted::Bool = true,
    preview::Bool = false,
    meta_data::Union{Vector{Dict{String, T}}, Nothing} = nothing
) where T <: Any

Load a DataMap instance with keys ids that map to data at all_paths, where every subvector refers the the paths of the data of a single dataset.

ModelWeights.Data.loadModelsFromCSV — Method

loadModelsFromCSV(
    path::String, col_models::String; col_variants::Union{String,Nothing}=nothing
)

Return a vector with models retrieved from csv file at path. If col_variants is provided, returned models are on level of model members (MODEL#variant, e.g. AWI-ESM#r1i1p1f1).

ModelWeights.Data.loadPreprocData — Method

loadPreprocData(
    paths::Vector{String},
    filename_format::Union{Symbol, String};
    sorted::Bool = true, 
    dtype::String = "undef",
    names::Vector{String} = Vector{String}(),
    meta_info::Union{Dict{String, T}, Nothing} = nothing
) where T <: Any

Return data loaded from paths as single YAXArray.

Each path points to a different model, i.e. the data for one dataset is loaded from multiple files and all datasets in paths must share the same dimensions. Which variable is loaded is inferred from the filenames (from paths), or if meta_info has key 'variable', the respective value is used.

ModelWeights.Data.memberIDFromFilenameMeta — Method

memberIDFromFilenameMeta(fn_meta::FilenameMeta)

Return a string of the form modelname#memberID[_grid] that identifies the corresponding model member.

For CMIP6 models the abbreviation of the grid is added to the model name.

ModelWeights.Data.memberIDsFromPaths — Method

memberIDsFromPaths(all_paths::Vector{String})

For every path in all_paths return a string of the form modelname#memberID[grid] that identifies the corresponding model member. Filenames must follow the CMIP-standard (projnamemipexpidvariable[_grid].nc).

For CMIP6 models the abbreviation of the grid is added to the model name.

ModelWeights.Data.mergeMetaDataFromMultipleFiles — Method

mergeMetaDataFromMultipleFiles(data::Vector{<:YAXArray})

Combine arrays in data into a single YAXArray with meta combined from all datasets into lists, with missing if key wasnt in a dataset.

ModelWeights.Data.metaDataChecksCMIP — Method

metaDataChecksCMIP(meta::Dict, path::String)

Check model names as retrieved from the metadata for potential inconsistencies wrt filename.

ModelWeights.Data.metaDataFromESMValToolRecipes — Method

metaDataFromESMValToolRecipes(
    base_path_configs::String; constraint::Union{Dict, Nothing} = nothing
)

Read variable, statistic, experiment and timerange/alias values from ESMValTool recipes stored at base_path_configs into a vector of Dictionaries storing the respective readoff values.

ModelWeights.Data.metaDataFromYAML — Method

metaDataFromYAML(content::Dict)

Return metadata of data specified in content possibly constrained by values in arg_constraint.

For constraints that are specified in content as well as in the arg_constraint argument, the values of the latter have precedence over the former. The constraints given in the argument arg_constraint are applied to EVERY dataset specified in the config file.

Arguments:

content: content of config yaml file specifying meta attributes and paths of data

ModelWeights.Data.modelDim — Method

modelDim(data::YAXArray)

Return the model dimension of data which must either be :model or :member.

If none is present, throw ArgumentError.

ModelWeights.Data.normalize — Method

normalize(data::Dict{String, <:Number})

Normalize values for every entry in data such that they sum up to 1. If remove_zero is true (default), the returned dictionary does not contain entries for which values were 0.

ModelWeights.Data.normalizeToYAX — Method

normalizeToYAX(data::Dict{String, <:Number})

Normalize values for every entry in data such that they sum up to 1 and return a YAXArray with dimension 'diagnostic' whose lookup names are the keys of data.

ModelWeights.Data.putAtModel! — Method

putAtModel!(data::YAXArray, dimension::Symbol, model::String, input)

ModelWeights.Data.readDataFromDisk — Method

readDataFromDisk(target_path::String; variable::String="")

Load data from target_path. If target_path ends with '.jld2', variable must be specified, otherwise data is assumed to be binary.

ModelWeights.Data.resolvePathsFromMetaData — Method

resolvePathsFromMetaData( 
    meta::MetaData,
    base_path::String,
    dir_per_var::Bool;
    constraint::Union{Dict, Nothing} = nothing
)

Return paths to data files for data specified in meta, possibly constraint by values in constraint. The paths were the data is stored is expected to follow the following structure (corresponding to the output from ESMValTool used for preprocessing the data):

base_path is the top-level directory. If dir_per_var is true, base_path is assumed to have a (or several) subdirectory for each climate variable with VAR as part of the subdirectory's name (e.g. _tas, cmip5tas, etc.). These subdirectories may be constraint by containing at least one of the values in constraint["base_subdirs"].

Let BASE refer to base_path, or respectively, to the subdirectories for the climate variables. Then the following structure is: BASE/preproc/meta.alias/meta.subdir. In ESMValTool, meta.alias corresponds to the (self-chosen) name under the section 'diagnostics' and meta.subdir to the (self-chosen) name under the section 'variables'.

The returned paths are the paths to all files within this directory, possibly constraint by the filenames containig at least one string in constraint["projects"] and respectively at least one string in constraint["models"].

ModelWeights.Data.setLookupsFromMemberToModel — Method

setLookupsFromMemberToModel(data::YAXArray, dim_names::Vector{String})

Change the lookup values for the dimension 'member' to refer to the models, i.e. they are not unique anymore. This is done in preparation to group the data by the different models.

Arguments:

data::YAXArray: has at least dimensions in dim_names.

dim_names::Vector{String}: names of dimensions to be changed, e.g. 'member',

'member1' (would be changed to 'model', 'model1').

ModelWeights.Data.sharedLevelMembers — Method

sharedLevelMembers(data::DataMap)

Return vector of model members that are shared across all entries in data.

ModelWeights.Data.sharedLevelModels — Method

sharedLevelModels(data::DataMap)

Return vector of models (on level of models, not members) that are shared across all entries in data.

ModelWeights.Data.sharedModels — Method

sharedModels(
    all_paths::Vector{Vector{String}},
    level_shared::Level,
    fn_format::Union{Symbol, String}
)

Arguments:

all_paths: every entry refers to the paths to data files for the respective dataset

ModelWeights.Data.subsetModelData — Method

subsetModelData(datamap::DataMap; level::Level=MEMBER_LEVEL)

For those datasets in datamap that specify data on the level level (i.e. have dimension :member or :model), return a new DataMap with subset of data s.t. the new datasets all have the same models (MODELLEVEL) or members (MEMBERLEVEL).

If no models are shared across datasets, return the input datamap.

ModelWeights.Data.subsetModelData — Method

subsetModelData(data::YAXArray, shared_models::Vector{String})

Return subset of data containing only data from models in shared_models.

Takes care of metadata.

Arguments:

data: must have dimension 'member' or 'model'

shared_models: models, which can either be on level of models or members of models

('modelname#memberID[_grid]').

ModelWeights.Data.summarizeMembers! — Method

summarizeMembers!(data::DataMap)

Set values for every dataset in data to the average across all members of each model.

ModelWeights.Data.summarizeMembers — Method

summarizeMembers(data::YAXArray; fn::Function=Statistics.mean)

For each model compute a summary statistic (default: mean) across all its members.

The returned YAXArray has dimension 'model'.

Arguments:

data::YAXArray: YAXArray; must have dimension 'member'

updateMeta::Bool: set true if the vectors in the metadata refer to

different models. Set to false if vectors refer to different variables.

fn::Function: Function to be applied on data

ModelWeights.Data.summarizeMembersMatrix — Method

summarizeMembersMatrix(data::AbstractArray, updateMeta::Bool; fn::Function=Statistics.mean)

Compute the average across all members of each model for each given variable for model to model data, e.g. distances between model pairs.

Arguments:

data: with at least dimensions 'member1', 'member2'

updateMeta: set true if the vectors in the metadata refer to different models.

Set to false if vectors refer to different variables for instance.

ModelWeights.Data.summarizeMembersVector — Method

summarizeMembersVector(data::YAXArray; fn::Function=Statistics.mean)

For each model compute a summary statistic (default: mean) across all its members. The returned YAXArray has dimension :model (instead of :member).

Arguments:

data::YAXArray: YAXArray; must have dimension 'member'

fn::Function: Function to be applied on data

ModelWeights.Data.summarizeMeta! — Method

summarizeMeta(meta::Dict, models::Vector{String})

If simplify is true, use single value when all remaining elements in a vector in meta are identical, otherwise just keep vector.

Arguments:

meta:

indices: indices to remain in vectors mapped to in meta.

ModelWeights.Data.summarizeMeta — Method

summarizeMeta(meta::Dict, models::Vector{String})

If simplify is true, use single value when all remaining elements in a vector in meta are identical, otherwise just keep vector.

Arguments:

meta:

indices: indices to remain in vectors mapped to in meta.

ModelWeights.Data.throwErrorIfDimMissing — Method

throwErrorIfDimMissing(data::YAXArray)

ModelWeights.Data.throwErrorIfDimMissing — Method

throwErrorIfDimMissing(data::YAXArray, dims::Vector{Symbol})

Throw ArgumentError if data does not have ALL dimensions in dims when include=:all (default), or if data does not have ANY dimension in dims.

ModelWeights.Data.uncertaintyRanges — Method

uncertaintyRanges(
    data::AbstractArray; w::Union{DimArray, Nothing}=nothing, quantiles=[0.167, 0.833]
)

Compute weighted quantiles of timeseries data.

Arguments:

data::AbstractArray: must have dimensions 'time', 'model'.

w::Union{DimArray, Nothing}=nothing: must have dimensions 'model' and 'weight', each

'weight'-vector must sum up to 1.

quantiles=[0.167, 0.833]: vector with two entries between 0 and 1 representing the

lower and upper bound in this order.

ModelWeights.Data.updateGroupedDataMetadata — Method

updateGroupedDataMetadata(meta::Dict, grouped_data::DimensionalData.DimGroupByArray)

Summarize vectors in meta, refering to different models (members), such that each vector only contains N entries (N=number of models (i.e. without unique members)).

If the metadata for members of a model differ across members, the respective entry in the vector will be a vector itself.

ModelWeights.Data.writeDataToDisk — Method

writeDataToDisk(data, target_path::String)

Save data as Julia obj if target_path has ending '.jld2', otherwise save as binary. If file at target_path already exists, timestamp is added if overwrite is false (default).

ModelWeights.Plots.convertKgsToSv! — Method

convertKgsToSv!(vec:YAXArray)

Convert data given in unit 'kg s-1' into Sverdrups (Sv).

ModelWeights.Plots.getClosestGridPoint — Method

getClosestGridPoint(location::Dict, longitudes::Vector, latitudes::Vector)

Find the grid point in grid defined by longitudes and latitudes that is closest to location.

Arguments

location::Dict: 'lon', 'lat' of position for which closest grid point is

returned, lon must be given from -180° to 180°

longitudes::Vector: grid longitudes measured from -180° to 180°

latitudes::Vector: grid latitudes measured from -90° to 90°

ModelWeights.Plots.latitude2NorthSouth — Method

latitude2NorthSouth(lat::Number)

Convert latitudes from -90° to 90° into 0° to 90° North/South.

ModelWeights.Plots.lon180to360 — Method

lon360to180(lon::Number)

Convert longitudes measured from -180° to 180° into 0° to 360° scale.

ModelWeights.Plots.lon360to180 — Method

lon360to180(lon::Number)

Convert longitudes measured from 0° to 360° into -180° to 180° scale.

ModelWeights.Plots.longitude2EastWest — Method

longitude2EastWest(lon::Number)

Convert longitudes from -180° to 180° into 0° to 180° East/West.

ModelWeights.Plots.makeSubplots — Method

makeSubplots

Arguments:

ModelWeights.Plots.plotAMOC — Method

plotAMOC(data::DimArray)

Plot AMOC strength for variable "amoc".

ModelWeights.Plots.plotDistances — Method

plotDistances(dists::AbstractArray, title::String; is_bar_plot::Bool=true)

Plot figure of distances for every combination of variable and diagnostic in dists.

Arguments:

dists: must have dimensions :variable, :diagnostic

title:

is_bar_plot:

ModelWeights.Plots.plotDistancesIndependence — Method

plotDistancesIndependence(distances::AbstractArray, dimname::String)

Arguments:

distances:

dimname:

ModelWeights.Plots.plotEnsembleSpread — Method

plotEnsembleSpread(data::AbstractArray, lon::Number, lat::Number)

Create figure with boxplots for each model in data that have several ensemble members.

ModelWeights.Plots.plotHistAtPos — Method

plotHistAtPos(data::AbstractArray, location::Dict)

Plot histogram of all data for a specific location.

Arguments:

data: dimensions 'lon' (from -180° to 180°), 'lat' (-90° to 90°)

location: must have keys 'name', 'lon', 'lat'

ModelWeights.Plots.plotTempGraph — Method

plotTempGraph

# TODO: change quantileLabels and uncertaintyRanges, the latter should contain 
# the quantile labels

ModelWeights.Plots.plotTimeseries! — Method

plotTimeseries!(ax::Axis, vals::AbstractArray;)

Plot timeseries of data vector vals.

ModelWeights.Plots.plotValsOnMap! — Method

plotValsOnMap!(fig::Figure, means::AbstractArray, title::String; colors=nothing, colorrange=nothing, highclip=(1,0,0), lowclip=(0,0,1), pos=(x=1, y=1), poslegend=nothing )

Plot contours of world with an overlayed heatmap of the input data.

Arguments:

nb_ticks: if nothing (default), just min lat/lon, max lat/lon and 0 are labeled.

pos::NamedTuple(x::Int,y::Int): position of plot in fig

ModelWeights.Plots.plotWeightContributions — Method

plotWeightContributions(independence::DimArray, performance::DimArray)

Plot performance against independence weights.

Arguments:

independence: (dims:model) normalized independence weights

performance: (dims:model) normalized performance weights

ModelWeights.Plots.plotWeights — Method

plotWeights(weights::YAXArray; title::String="", sort_by::String="")

Plot all weights sorted by weight dimension sort_by of weights if given, otherwise, sorted by first value in weight dimension.

Equal weights line is added to plot too.

Arguments:

weights:

title:

sort_by: value of weight dimension according to which data is sorted

ModelWeights.Plots.sortLongitudesWest2East — Method

sortLongitudesWest2East(data::AbstractArray)

Arrange 'data' such that western latitudes come first, then eastern latitudes.

ModelWeights.Timeseries.addFilteredTimeseries! — Method

addFilteredTimeseries!(
    data::DataMap, 
    start_year::Number, 
    end_year::Number;
    ids_ts::Vector{String}=Vector{String}(),
    new_alias::String,
    only_models_non_missing_vals::Bool = true
)

ModelWeights.Timeseries.getUncertaintyRanges — Method

getUncertaintyRanges(
    data::AbstractArray; w::Union{DimArray, Nothing}=nothing, quantiles=[0.167, 0.833]
)

Compute weighted quantiles of timeseries data.

Arguments:

data::AbstractArray: must have dimensions 'time', 'model'.

w::Union{DimArray, Nothing}=nothing: must have dimensions 'model' and 'weight', each

'weight'-vector must sum up to 1.

quantiles=[0.167, 0.833]: vector with two entries between 0 and 1 representing the

lower and upper bound in this order.

ModelWeights.Timeseries.linearTrend — Method

linearTrend(data::YAXArray; full_predictions::Bool=true)

Compute linear trend as ordinary least squares for timeseries data.

In the metadata of the returned array '-TREND' (if full_predictions=false) or '-TREND-pred' (if full_predictions=true) is added to the original metadata entry of data for 'statistic' and the new metadata entry for 'id' is adapted accordingly.

Arguments:

data::YAXArray: must have dimension :time.

full_predictions::Bool: if false, return only slope, otherwise return predicted value.

ModelWeights.Weights.addLikelihoodWeights! — Method

addLikelihoodWeights!(
    data::ClimateData, 
    values::Vector{<:Number}, 
    models::Vector{String}, 
    distr::Distribution,
    id::String
)

Compute the likelihood weights for values with respect to the distribution distr and add computed weights to weights Array of data with name id. If id is already present, nothing is added.

ModelWeights.Weights.applyWeights — Method

applyWeights(model_data::YAXArray, weights::YAXArray)

Compute the weighted average of model data data with given weights.

If weights were computed for a superset of the models in data, they are normalized and applied to the subset. Only weights per model (not members) are considered for now, in the future, members should be considered too.

Arguments:

model_data::YAXArray: model predictions. If given for model members, the predictions

of each model are considered the average value of all members of the respective model.

weights::YAXArray: if given for each member of a model, these will be summed up to

yield one value per model.

ModelWeights.Weights.climwipWeights — Method

climwipWeights(
    dists_indep_all::YAXArray, 
    dists_perform_all::YAXArray,
    config::ConfigWeights;
    suffix_name::Stirng = "climwip"
)

Compute weight for each model in multi-model ensemble according to approach from Brunner, Lukas, Angeline G. Pendergrass, Flavio Lehner, Anna L. Merrifield, Ruth Lorenz, and Reto Knutti. “Reduced Global Warming from CMIP6 Projections When Weighting Models by Performance and Independence.” Earth System Dynamics 11, no. 4 (November 13, 2020): 995–1012. https://doi.org/10.5194/esd-11-995-2020.

Arguments:

dists_indep_all::YAXArray: RMSEs between pairs of models for all

combinations of variables and diagnostics; has dimensions 'member1', 'member2', 'variable', 'diagnostic'.

dists_perform_all::YAXArray: RMSEs between model and observational data for all

combinations of variables and diagnostics; has dimensions 'member', 'variable', 'diagnostic'.

config::ConfigWeights: Parameters specifiying the relative contributions of each

combination of variable and diagnostic.

suffix_name::String: added to name of each type of weights, s.t. names are: wP-suffix,

wI-suffix, combined-suffix (default 'climwip').

ModelWeights.Weights.distributeWeightsAcrossMembers — Method

distributeWeightsAcrossMembers(weights::YAXArray, members::Vector{String})

Equally distribute weight for each model over its members.

ModelWeights.Weights.equalWeights — Method

equalWeights(data::YAXArray; use_members::Bool=true)

Create a weight vector, with equal weight for each MODEL. Distribute weight across model members if dimension=:member and usemembers is true. If usemember is false, each model member is considered as standalone model and all receive the same weight.

ModelWeights.Weights.independenceWeights — Method

independenceWeights(Sij::YAXArray, sigmaS::Number)

Compute the independence weight for each model i with generalized distances between pairs of models, Sij.

\[w^{I}_i = a_0 \cdot \frac{1}{1 + \sum_{j \ne i} e^{-\left( \frac{S_{ij}}{\sigma_S} \right)^2}}\]

Arguments:

Sij: generalized distances for each pair of models with dimensions 'model1', 'model2'

sigmaS: free shape parameter for impact of independence weights

ModelWeights.Weights.loadWeightsAsDimArray — Method

loadWeightsAsDimArray(data::NCDataset, key_weights::String)

Arguments:

data: NCDataset containing weights, which have a single dimension

key_weights: name of weights to load; 'wP' (performance weights), 'wI'

(independence weights), 'w' (overall weights)

ModelWeights.Weights.performanceWeights — Method

performanceWeights(Di::AbstractArray, sigmaD::Number)

Compute the ClimWIP performance weight for each model i with generalized distances Di.

\[w^{P}_i = a_0 \cdot \frac{e^{-(\frac{D_i}{\sigma_D})^2}}{N}\]

Arguments:

Di::YAXArray: generalized distances Di for each model

sigmaD::Number: free shape parameter for impact of performance weights

ModelWeights.Weights.saveWeightsAsNCFile — Method

saveWeightsAsNCFile(weights::ClimWIP; target_path::String)

TODO: needs to be updated, deprecated

Arguments:

weights: ClimWIP object to be saved.

target_path: Path to where weights shall be stored.

ModelWeights.Weights.weightedAvg — Method

weightedAvg(
    data::YAXArray; 
    weights::Union{DimArray, Nothing}=nothing, 
    use_members_equal_weights::Bool=true
)

Compute the average values for each (lon,lat) grid point in data, weighted by weights.

If no weight vector is provided, unweighted average is computed.

Arguments:

data::YAXArray: must have dimension 'model' or 'member'

weights::Union{YAXArray, Nothing}=nothing: weights for models or individual members

use_members_equal_weights::Bool: if weights is nothing, all models receive

equal weight. If use_members_equal_weights is true, the number of members per model is taken into account, s.t. each model receives equal weight, which is distributed among the respective members.

Index

ModelWeights.Data.addDiagnostic!

ModelWeights.Data.addDiagnostic!

ModelWeights.Data.alignPhysics

ModelWeights.Data.anomalies

ModelWeights.Data.approxAreaWeights

ModelWeights.Data.areaWeightedRMSE

ModelWeights.Data.checkConstraint

ModelWeights.Data.combineAll

ModelWeights.Data.combineModelsFromMultipleFiles

ModelWeights.Data.constrainMetaData!

ModelWeights.Data.defineDataMap

ModelWeights.Data.defineDataMap

ModelWeights.Data.defineDataMap

ModelWeights.Data.distancesData

ModelWeights.Data.distancesData

ModelWeights.Data.distancesModels

ModelWeights.Data.filterPathsSharedModels

ModelWeights.Data.filterPathsSharedModels

ModelWeights.Data.generalizedDistances

ModelWeights.Data.getAtModel

ModelWeights.Data.getByIdxModel

ModelWeights.Data.getCMIPModelsKey

ModelWeights.Data.globalMeans

ModelWeights.Data.individuatePath

ModelWeights.Data.interpolatedWeightedQuantiles

ModelWeights.Data.joinDicts

ModelWeights.Data.kelvinToCelsius

ModelWeights.Data.kelvinToCelsius!

ModelWeights.Data.loadDataFromESMValToolRecipes

ModelWeights.Data.loadDataFromYAML

ModelWeights.Data.loadDataMapCore

ModelWeights.Data.loadModelsFromCSV

ModelWeights.Data.loadPreprocData

ModelWeights.Data.memberIDFromFilenameMeta

ModelWeights.Data.memberIDsFromPaths

ModelWeights.Data.mergeMetaDataFromMultipleFiles

ModelWeights.Data.metaDataChecksCMIP

ModelWeights.Data.metaDataFromESMValToolRecipes

ModelWeights.Data.metaDataFromYAML

ModelWeights.Data.modelDim

ModelWeights.Data.normalize

ModelWeights.Data.normalizeToYAX

ModelWeights.Data.putAtModel!

ModelWeights.Data.readDataFromDisk

ModelWeights.Data.resolvePathsFromMetaData

ModelWeights.Data.setLookupsFromMemberToModel

ModelWeights.Data.sharedLevelMembers

ModelWeights.Data.sharedLevelModels

ModelWeights.Data.sharedModels

ModelWeights.Data.subsetModelData

ModelWeights.Data.subsetModelData

ModelWeights.Data.summarizeMembers

ModelWeights.Data.summarizeMembers!

ModelWeights.Data.summarizeMembersMatrix

ModelWeights.Data.summarizeMembersVector

ModelWeights.Data.summarizeMeta

ModelWeights.Data.summarizeMeta!

ModelWeights.Data.throwErrorIfDimMissing

ModelWeights.Data.throwErrorIfDimMissing

ModelWeights.Data.uncertaintyRanges

ModelWeights.Data.updateGroupedDataMetadata

ModelWeights.Data.writeDataToDisk

ModelWeights.Plots.convertKgsToSv!

ModelWeights.Plots.getClosestGridPoint

ModelWeights.Plots.latitude2NorthSouth

ModelWeights.Plots.lon180to360

ModelWeights.Plots.lon360to180

ModelWeights.Plots.longitude2EastWest

ModelWeights.Plots.makeSubplots

ModelWeights.Plots.plotAMOC

ModelWeights.Plots.plotDistances

ModelWeights.Plots.plotDistancesIndependence

ModelWeights.Plots.plotEnsembleSpread

ModelWeights.Plots.plotHistAtPos

ModelWeights.Plots.plotTempGraph

ModelWeights.Plots.plotTimeseries!

ModelWeights.Plots.plotValsOnMap!

ModelWeights.Plots.plotWeightContributions

ModelWeights.Plots.plotWeights

ModelWeights.Plots.sortLongitudesWest2East

ModelWeights.Timeseries.addFilteredTimeseries!

ModelWeights.Timeseries.getUncertaintyRanges

ModelWeights.Timeseries.linearTrend

ModelWeights.Weights.addLikelihoodWeights!

ModelWeights.Weights.applyWeights

ModelWeights.Weights.climwipWeights

ModelWeights.Weights.distributeWeightsAcrossMembers

ModelWeights.Weights.equalWeights

ModelWeights.Weights.independenceWeights

ModelWeights.Weights.loadWeightsAsDimArray

ModelWeights.Weights.performanceWeights

ModelWeights.Weights.saveWeightsAsNCFile

ModelWeights.Weights.weightedAvg