Usage

Climwip example from Brunner et al. (2020)

Weights for models of lgm experiment

Load model data for experiment lgm and historical in one run from new config file

TODO: show that we get the same outfit different approaches!

Getting started

How to load data

How to compute weights

Usage

DataMap Object

Weight Objects

Application of weights

Functions

Index

References

Requirements

Data structure

Config files

Computing weights

Area-weighted RMSE

Generalized distances

Computation of overall weight for model $i$

Computation of performance/independence weights only

The purpose of this package is twofold: on the one hand, we built it to handle preprocessed Earth System data easily within Julia and, on the other hand, to compute weights for different sets of Earth System Models following the approach from Brunner et al (2020).

The main objects that the package provides are: DataMap and two weight objects, ConfigWeights and Weights.

DataMap Object

Each Data object refers to some summary data (e.g. climatological average) of a particular experiment for the full time period or part of it.

DataMap is a type alias for a dictionary mapping from Strings to YAXArrays. The metadata of each dataset ($datamap[id].properties$) contains the metadata loaded from the original files as well as some additional information that we add:

_variable

_statistic

_alias

_timerange

_paths

_id: concatenates variable, statistic and alias into a single String

separated by underscores

Weight Objects

ConfigWeights

import ModelWeights as mw

# load model and observational data

config_weights = mw.ConfigWeights(
    performance = Dict("tas_CLIM" => 1, "tos_CLIM" => 1),
    independence = Dict("tas_CLIM" => 1, "tos_CLIM" => 1),
    sigma_performance = 0.5,
    sigma_independence = 0.5,
    alias_ref_perform_weights = "historical",
    alias_ref_indep_weights = "historical",
    target_path = ""
);

Weights

weights = mw.computeWeights(model_data_historical, obs_data, config_weights);

The Weights object has several fields that each store a DimensionalData.DimArray. Some of them refer to normalized weights, others refer to the distances on the basis of which the weights were computed.

performance_distances and independence_distances contain the distances

between models and data (performance) or between models and models (independence) for every combination of variable and diagnostic/statistic.

Di and Sij are the generalized distances, i.e. weighted average of all

distances across variables and diagnostic. So, $D_i$ is a vector with size 1xn where n is the number of models (not on level of members!) and $S_{ij}$ is a matrix of size n x n.

wP and wI respectively store the normalized performance/independence weights,

both have size 1xn (n: number of models).

w is the overall weight vector of length n.

Application of weights

TODO: add example of how weights are applied to data

Functions

ModelWeights.addAnomalies! — Method

addAnomalies!(
    datamap::DataMap id_data::String, id_ref::String; stats::String="ANOM"
)

Add entry to datamap with difference between datamap at id_data and id_ref.

The id of the original data and of the reference data is added to the metadata.

ModelWeights.addAnomaliesGM! — Method

addAnomaliesGM!(data::DataMap, ids_data::Vector{String})

Compute anomalies of datasets in data with ids in ids_data with respect to their global means.

ModelWeights.addDiagnostic! — Method

addDiagnostic!(datamap::DataMap, fn::Function, args...; kwargs...)

Compute a diagnostic by running fn with positional arguments args and keyword arguments kwargs and add result to datamap at the id of the computed result.

ModelWeights.addGlobalMeans! — Method

addGlobalMeans!(data::DataMap; ids::Union{Vector{String},Nothing})

Compute global means for datasets in data with ids.

If ids is not specified, compute global means for all datasets in data.

ModelWeights.addLinearTrend! — Method

addLinearTrend!(
    data::DataMap;     
    statistic::String="CLIM-ann", 
    ids_ts::Vector{String}=Vector{String}(), 
    full_predictions::Bool=true
)

Add computed linear trend (computeLinearTrend) for all datasets in data with an id in ids_ts, or if ids_ts is not given for all datasets with statistic (as specified in metadata _statistic), where the default is the annual climatologies (CLIM-ann).

ModelWeights.addTempSTD! — Method

addTempSTD!(data::DataMap; statistic::String="CLIM-ann")

Compute temporal standard deviation (computeTempSTD) for every dataset with statistic in its id (default: CLIM-ann) and add the result to data.

The id of computed temporal standard deviation is the same as before, but with statistic replaced by 'TREND'.

ModelWeights.alignPhysics — Method

alignPhysics(
    data::DataMap,
    members::Vector{String}, 
    subset_shared::Union{LEVEL, Nothing} = nothing)
)

Return new DataMap with only the models retained that share the same physics as the respective model's members in members.

If subset_shared is set, resulting DataMap is subset accordingly.

ModelWeights.applyDataConstraints! — Method

applyDataConstraints!(meta_attributes::Vector{Dict{String, Any}}, constraint::Dict)

Subset entries in meta_attributes so that only those with properties specified in constraint remain.

Arguments

constraint::Dict: Mapping to vector specifiying the properties of which at

least one must be present for an id to be retained.

ModelWeights.applyModelConstraints — Method

applyModelConstraints(
    path_model_data::String, model_constraints::Vector{String}
)

Return true if constraints in model_constraints are fulfilled, i.e. if the given path_model_data contains any model from model_constraints, false otherwise.

Arguments:

model_constraints: strings that may contain only model name, e.g. 'MPI-ESM-P',

or modelname and member id, e.g. 'MPI-ESM-P#r1i1p2' or model name, member id and grid, e.g. 'MPI-ESM-P#r1i1p2gn'.

ModelWeights.applyWeights — Method

applyWeights(model_data::YAXArray, weights::YAXArray)

Compute the weighted average of model data 'data' with given weights 'weights'.

If weights were computed for a superset of the models in 'data', they are normalized and applied to the subset. Only weights per model (not members) are considered for now, in the future, members should be considered too.

Arguments:

model_data::YAXArray: model predictions. If given for model members, the predictions

of each model are considered the average value of all members of the respective model.

weights::YAXArray: if given for each member of a model, these will be summed up to

yield one value per model.

ModelWeights.approxAreaWeights — Method

approxAreaWeights(latitudes::Vector{<:Number})

Create a YAXArray with the cosine of latitudes which approximates the cell area on the respective latitude.

ModelWeights.areaWeightedRMSE — Method

areaWeightedRMSE(m1::AbstractArray, m2::AbstractArray, mask::AbstractArray)

Compute the area weighted (approximated by cosine of latitudes in radians) root mean squared error between m1 and m2.

Arguments:

m1: must have dimensions 'lon', 'lat'.

m2: must have dimensions 'lon', 'lat'.

mask: has values 0,1. Locations where mask is 1 get a weight of 0.

ModelWeights.averageEnsembleMembersMatrix — Method

averageEnsembleMembersMatrix(data::AbstractArray, updateMeta::Bool)

Compute the average across all members of each model for each given variable for model to model data, e.g. distances between model pairs.

Arguments:

data: with at least dimensions 'member1', 'member2'

updateMeta: set true if the vectors in the metadata refer to different models.

Set to false if vectors refer to different variables for instance.

ModelWeights.buildCMIP5EnsembleMember — Method

buildCMIP5EnsembleMember(
    realizations::Vector, initializations::Vector, physics::Vector
)

Concatenate model settings to build ripf-abbreviations for CMIP5 models which do not have it in their metadata. Return a vector of strings.

ModelWeights.buildPathsForMetaAttrib — Method

buildPathsForMetaAttrib(
    base_path::String, 
    attrib::MetaAttrib,
    dir_per_var::Bool;
    subdir_constraints::Union{Vector{String}, Nothing}=nothing
)

Return paths to data specified by attribs.

Assumed data structure: BASE/preproc/ALIAS/VARSTAT where BASE=`basepath, ALIAS=attribs["alias"]`,VAR=attribs["variable"] and STAT=attribs["statistic"]. If `dirpervaris true, BASE is path to any subdirectory ofbasepathwhose name contains _VAR. Ifsubdirconstraintsis given, the subdirectory's name must further contain at least one entry insubdirconstraints`.

Arguments:

base_path: base directory of stored data specified in attrib.

attribs: meta attributes of data. Must have keys: 'variable', 'statistic', '_alias'.

dir_per_var: true if data of each climate variable is stored in a seperate directory.

subdir_constraints: if given, paths must contain ANY of the given elements. Existing paths that don't are ignored.

ModelWeights.buildPathsToDataFiles — Method

buildPathsToDataFiles(
    path_data::String,
    is_model_data::Bool;
    model_constraints::Vector{String} = Vector{String}(),
    project_constraints::Vector{String} = Vector{String}()
)

Build vector of strings containing paths to data files in path_data that were not filtered out by model_constraints or project_constraints.

Arguments:

model_constraints::Vector{String}:

project_constraints::Vector{String}:

ModelWeights.computeDistancesAllDiagnostics — Method

computeDistancesAllDiagnostics(
    model_data::DataMap, 
    obs_data::Union{Nothing, DataMap}, 
    config::Dict{String, Number},
    ref_period_alias::String,
    forPerformance::Bool    
)

Compute RMSEs between models and observations or between predictions of models for all variables and diagnostics for which weights are specified in config.

Arguments:

config::Dict{String, Number}: mapping from 'VARIABLE_DIAGNOSTIC' to

respective weight.

for_performance::Bool: true for distances between models and observations,

false for distances between model predictions.

ModelWeights.computeGeneralizedDistances — Method

computeGeneralizedDistances(
    distances_all::YAXArray, weights::DimArray, for_performance::Bool

)

For every variable in distances_all, compute the weighted sum of all diagnostics.

ModelWeights.computeGlobalMeans — Method

computeGlobalMeans(data::AbtractArray)

Return a YAXArray with area-weighted global means for each model in data.

Missing data is accounted for in the area-weights.

Arguments:

data::YAXArray: must have dimensions 'lon' and 'lat'.

ModelWeights.computeInterpolatedWeightedQuantiles — Method

computeInterpolatedWeightedQuantiles(
    quantiles::Vector{<:Number}, vals::Vector; weights=nothing    
)

This implementation follows the one used by Brunner et al.

ModelWeights.computeLinearTrend — Method

computeLinearTrend(data::YAXArray; full_predictions::Bool=true)

Compute linear trend as ordinary least squares for timeseries data.

Arguments:

data::YAXArray: must have dimension :time.

full_predictions::Bool: if false, return only slope, otherwise return

predicted value.

ModelWeights.computeModelDataRMSE — Method

computeModelDataRMSE(
    model_data::DataMap, obs_data::DataMap, config::ConfigWeights    
)

Compute the Root-mean-squared-error between model_data and obs_data for each combination of variable and diagnostic for which weights > 0 are specified in config.

ModelWeights.computeModelModelRMSE — Method

 computeModelModelRMSE(model_data::DataMap, config::ConfigWeights)

Compute the Root-mean-squared-error between pairs of models in model_data for each combination of variable and diagnostic for which weights > 0 are specified in config.

ModelWeights.computeTempSTD — Method

computeTempSTD(data::YAXArray, trend::YAXArray)

Compute the standard deviation of the temporal detrended timeseries data as STD^{t2}t=t1 (Xl^t) - X_l^{TREND} where l is a (lon, lat)-position and TREND is the prediction of the ordinary least squares fit of the data (i.e. trend must have the same size as data).

ModelWeights.computeWeightedAvg — Method

computeWeightedAvg(
    data::YAXArray; 
    weights::Union{DimArray, Nothing}=nothing,
    use_members_equal_weights::Bool=true
)

Compute the average values for each (lon,lat) grid point in 'data', weighted by weights 'weights'. If no weight vector is provided, unweighted average is computed.

Arguments:

data: must have dimension 'model' or 'member'

weights: weights for models or individual members

use_members_equal_weights: if weights is nothing, all models receive

equal weight. If use_members_equal_weights is true, the number of members per model is taken into account, s.t. each model receives equal weight, which is distributed among the respective members.

ModelWeights.computeWeights — Method

computeWeights(
    dists_indep_all::YAXArray, 
    dists_perform_all::YAXArray,
    config::ConfigWeights
)

Compute weight for each model in multi-model ensemble according to approach from Brunner, Lukas, Angeline G. Pendergrass, Flavio Lehner, Anna L. Merrifield, Ruth Lorenz, and Reto Knutti. “Reduced Global Warming from CMIP6 Projections When Weighting Models by Performance and Independence.” Earth System Dynamics 11, no. 4 (November 13, 2020): 995–1012. https://doi.org/10.5194/esd-11-995-2020.

Arguments:

dists_indep_all::YAXArray: RMSEs between pairs of models for all

combinations of variables and diagnostics; has dimensions 'member1', 'member2', 'variable', 'diagnostic'.

dists_perform_all::YAXArray: RMSEs between model and observational

data for all combinations variables and diagnostics; has dimensions 'member', 'variable', 'diagnostic'.

config::ConfigWeights: Parameters specifiying the relative contributions

of each combination of variable and diagnostic.

ModelWeights.convertKgsToSv! — Method

convertKgsToSv!(vec:YAXArray)

Convert data given in unit 'kg s-1' into Sverdrups (Sv).

ModelWeights.distributeWeightsAcrossMembers — Method

distributeWeightsAcrossMembers(weights::YAXArray, members::Vector{String})

Equally distribute weight for each model over its members.

ModelWeights.filterModels — Method

filterModels(data::YAXArray, remaining_models::Vector{String})

Remove models from data that aren't in remaining_models and adapt the metadata of data accordingly such that 'membernames' and 'modelnames' corresponds to the remaining data.

ModelWeights.filterPaths — Method

filterPaths(paths::Vector{String}, shared_models::Vector{String})

Return new vector with paths in paths that point to models in shared_models.

Arguments:

shared_models: have form 'modelname#memberID[_grid]' on level of model

members or just 'modelname' on level of models.

ModelWeights.filterPathsSharedModels! — Method

filterPathsSharedModels!(
    meta_data::Dict{String, Dict{String, Any}},
    subset_shared::Union{LEVEL, Nothing}
)

In particular important before loading data if data should be subset

ModelWeights.filterTimeseries — Method

filterTimeseries(
    data_all::DataMap, 
    start_year::Number, 
    end_year::Number;
    new_alias::String="",
    ids_ts::Vector{String}=Vector{String}(),
    only_models_non_missing_vals::Bool = true
)

ModelWeights.getAliasAsTimerange — Method

getAliasAsTimerange(meta_attribs::Vector{MetaAttrib}, alias::String)

Translate each alias in meta_attribsto corresponding timerange.

ModelWeights.getAllCombinations — Method

getAllCombinations(v...)

Generate vector of strings with all possible combinations of input vectors, where each combination consists of one element from each input vector, concatenated as a string with underscores separating the elements.

Arguments

v...: A variable number of input vectors.

Example

julia> ModelWeights.getAllCombinations(["tos", "tas"], ["CLIM"])
2-element Vector{String}:
 "tos_CLIM"
 "tas_CLIM"

ModelWeights.getAtModel — Method

getAtModel(data::AbstractArray, dimension::Symbol, model::String)

Return data where dimension (member or model) has value model.

ModelWeights.getByIdxModel — Method

getByIdxModel(data::AbstractArray, dimension::Symbol, indices::Vector)

Return data where dimension (member or model) has value model.

ModelWeights.getCMIPModelsKey — Method

getCMIPModelsKey(meta::Dict)

Return the respective key to retrieve model names in CMIP6 ('sourceid') and CMIP5 ('modelid') data.

If both keys are present, 'source_id' used in CMIP6 models is returned, if none is present, throw ArgumentError.

ModelWeights.getClosestGridPoint — Method

getClosestGridPoint(location::Dict, longitudes::Vector, latitudes::Vector)

Find the grid point in grid defined by longitudes and latitudes that is closest to location.

Arguments

location::Dict: 'lon', 'lat' of position for which closest grid point is

returned, lon must be given from -180° to 180°

longitudes::Vector: grid longitudes measured from -180° to 180°

latitudes::Vector: grid latitudes measured from -90° to 90°

ModelWeights.getDimsModel — Method

getDimsModel(data::AbstractArray)

Return Tuple with the model dimension of data ('model' or 'member') as Symbol at position 1 and the respetive dimension names at position 2.

ModelWeights.getMemberIDsFromPaths — Method

getMemberIDsFromPaths(all_paths::Vector{String})

For every path in all_paths return a string of the form modelname#memberID[_grid] that identifies the corresponding model member.

The abbreviation of the grid is added to the model name for CMIP6 models.

ModelWeights.getMetaAttributesFromESMValToolConfigs — Method

getMetaAttributesFromESMValToolConfigs(
    base_path_configs::String;
    constraint::Union{Dict, Nothing} = nothing

)

Read variable, statistic, experiment and timerange/alias values from ESMValTool recipes stored at base_path_configs into a vector of Dictionaries storing the respective readoff values.

If constraint is given, a combination of 'statistic', 'experiment', etc. is only retained if each of the respective readoff values aligns with the the specified constraints if provided.

ModelWeights.getMetaDataFromYAML — Method

getMetaDataFromYAML(
    path_config::String,
    is_model_data::Bool
    subset::Union{Dict, Nothing} = nothing
)

Load data as specified in config file located at path_config. For constraints that are specified in the config file as well as in the arg_constraint argument, the values of the latter have precedence over the former. The constraints given in the argument subset are applied to EVERY dataset specified in the config file.

Arguments:

content: content of config yaml file specifying meta attributes and paths of data

is_model_data: true for model data, false for observational data

arg_constraint: TODO

ModelWeights.getModelDataDist — Method

getModelDataDist(models::YAXArray, observations::AbstractArray)

Compute the distance as the area-weighted RMSE between model predictions and observations.

ModelWeights.getModelDistances — Method

getModelDistances(modelData::YAXArray)

Compute the area weighted root mean squared error between model predictions for each pair of models.

Arguments:

modelData::YAXArray: must have dimensions 'lon', 'lat', 'model' and

contains the data for a single climate variable.

ModelWeights.getModelLogLikelihoods — Method

getModelLogLikelihoods(model_data::YAXArray, distr::Distribution)

Compute the log likelihood of model_data to originate from distr.

ModelWeights.getPathsToData — Method

getPathsToData(
    attribs::Dict{String, Any},
    base_path_data::String,
    dir_per_var::Bool,
    is_model_data::Bool;
    constraint::Union{Dict, Nothing}=nothing
)

Return the paths to the data files in base_path_data taking into account constraint.

ModelWeights.getRefPeriodAsTimerangeAndAlias — Method

getRefPeriodAsTimerangeAndAlias(
    meta_attribs::Vector{MetaAttrib}, ref_period::String
)

Return timerange and alias corresponding to ref_period which can be either be a timerange or an alias.

ModelWeights.getSharedModelsFromPaths — Method

getSharedModelsFromPaths(
    all_paths::Vector{Vector{String}}, all_models::Vector{String}
)

Return vector with those models in 'allmodels' for which a path is given in every subvector of 'metadata'.

ModelWeights.getTimerangeAsAlias — Method

getTimerangeAsAlias(meta_attribs::Vector{MetaAttrib}, timerange::String)

Translate given timerange to corresponding alias in data_ids.

ModelWeights.getUncertaintyRanges — Method

getUncertaintyRanges(
    data::AbstractArray; 
    w::Union{DimArray, Nothing}=nothing, 
    quantiles=[0.167, 0.833]}
)

Compute weighted quantiles of data.

Arguments:

data::AbstractArray: must have dimensions 'time', 'model'.

w::Union{DimArray, Nothing}=nothing: must have dimension 'model' and sum up to 1.

quantiles=[0.167, 0.833]: vector with two entries between 0 and 1

representing the lower and upper bound in this order.

ModelWeights.getUniqueMemberIds — Method

getUniqueMemberIds(meta::Dict, model_names::Vector{String})

Create a vector that contains for every model in model_names the unique ids of its members, each with the model name followed by '#' as prefix .

The unique member ids correspond to the variant labels of CMIP6 models, e.g. r1i1p1f1.

Arguments:

meta::Dict: For CMIP5 data, must have keys: 'mip_era', 'realization',

'initializationmethod', 'physicsversion'. For CMIP6 data must have keys: 'variantlabel', 'gridlabel'.

model_names::Vector{String}: Vector of strings containing model names for

every model member, i.e. length is sum of the number of members over all models.

ModelWeights.get_optional_fields_config — Method

get_optional_fields_config(ds::Dict, timerange_aliases_dict::Dict)

Fill optional fields for a dataset in config file with default values. Returned timeranges and aliases have the same length and correspond to one another.

Arguments:

ds:

timerange_aliases_dict:

ModelWeights.independenceParts — Method

independenceParts(generalizedDistances::AbstractArray, sigmaS::Number)

Compute the independence part (denominator) of the overall weight for each model.

Arguments:

generalizedDistances: contains generalized distances S_{i,j} for each model

(i.e. not model members) pair has two dimensions, 'model1' and 'model2'

sigmaS: free model parameter for impact of independence weights

ModelWeights.isValidDataAndWeightInput — Method

isValidDataAndWeightInput(
    data::DataMap, keys_weights::Vector{String}, ref_period_alias::String
)

Check that there is data in data for all keys_weights containing weights balancing the different diagnostics for the given reference period ref_period_alias.

ModelWeights.joinDataMaps — Method

joinDataMaps(v::DataMap...)

ModelWeights.kelvinToCelsius! — Method

kelvinToCelsius!(datamap::DataMap)

Modify entries of datamap such that all data is given in Degree Celsius (instead) of Kelvin.

Arguments:

datamap:

ModelWeights.kelvinToCelsius — Method

kelvinToCelsius(data::AbstractArray)

Return a copy of data with values given in Kelvin covnerted into Degree Celsius.

Arguments:

data:

ModelWeights.latitude2NorthSouth — Method

latitude2NorthSouth(lat::Number)

Convert latitudes from -90° to 90° into 0° to 90° North/South.

ModelWeights.loadDataFromESMValToolRecipes — Method

loadDataFromESMValToolRecipes(
    path_data::String,
    path_recipes::String;
    dir_per_var::Bool=true,
    is_model_data::Bool=true,
    subset::Union{Dict, Nothing}=nothing,
    preview::Bool=false
)

Loads the data from the config files (ESMValTool recipes) located at 'path_recipes'. For each variable, experiment, statistic and timerange (alias), an instance of Data is created.

Arguments:

path_data: path to location where preprocessed data is stored; if

dirpervar is true, paths to directories that contain one or more subdirectories that each contains a directory 'preproc' with the preprocessed data. If dirpervar is false, path_data is path to directory that directly contains the 'preproc' subdirectory.

path_recipes: path to directory that contains one or ESMValTool recipes

used as config files.

dir_per_var: if true (default), directory at path_data has subdirectories, one for

each variable (they must end with _ and the name of the variable), otherwise data_path points to the directory that has a subdirectory 'preproc'.

is_model_data: set true for model data, false for observational data

subset: specifies the properties of the subset of data to be loaded.

The following keys are considered: models (used to load only specific set of models or members of models), projects (used to load only data from a given set of projects, e.g. for loading only CMIP5-data), timeranges and aliases. (super set is loaded, i.e. data that corresponds to either a given timerange or a given alias will be loaded), variables, statistics, subdirs and subset_shared . (if set to MEMBER/MODEL only data loaded from model members/models shared across all experiments and variables is loaded).

preview: used to pre-check which data will be loaded before actually loading

it.

ModelWeights.loadDataFromMetadata — Method

loadDataFromMetadata(
    meta_data::Dict{String, Dict{String, Any}}, 
    is_model_data::Bool;
    subset::Union{Dict,Nothing} = nothing
)

Load a DataMap-instance that contains the data specified in meta_data possibly constraint by values in subset.

# Arguments:

- meta_data::Dict: TODO: add required keys!

- is_model_data::Bool: true for model data (default), false for observational data.

- subset: specifies the properties of the subset of data to be loaded. These

have to apply to each loaded dataset specified in the config yaml file.

The following keys are considered: models (used to load only specific set of models

or members of models), projects (used to load only data from a given set of

projects, e.g. for loading only CMIP5-data), timeranges and aliases

(super set is loaded, i.e. all data that corresponds to either a given timerange or

a given alias will be loaded), variables, statistics, subdirs, subset_shared

(if set to MEMBER/MODEL only data loaded from model members/models shared across all

datasets is loaded) and dir_per_var.

ModelWeights.loadDataFromYAML — Method

loadDataFromYAML(
    path_config::String;
    is_model_data::Bool = true,
    subset::Union{Dict,Nothing} = nothing,
    preview::Bool = false
)

Load a DataMap-instance that contains the data specified in yaml file at path_config, potentially constraint by values in subset.

Arguments:

preview::Bool: if true (default: false), return metadata without actually

loading any data.

ModelWeights.loadPreprocData — Method

loadPreprocData(meta::Dict{String, Any}, is_model_data::Bool)

Create a tuple with a vector of YAXArrays for data specified in meta and a Dictionary containing the metadata of all loaded data.

Load the data from paths specified in meta.paths and create a meta dictionary that contains the metadata keys from every loaded dataset. Each key maps to a vector of values, one for each loaded dataset, which is set to missing if that key hadn't been present in this datasets own metadata.

ModelWeights.loadWeightsAsDimArray — Method

loadWeightsAsDimArray(data::NCDataset, key_weights::String)

Arguments:

data: NCDataset containing weights, which have a single dimension

key_weights: name of weights to load; 'wP' (performance weights), 'wI'

(independence weights), 'w' (overall weights)

ModelWeights.lon180to360 — Method

lon360to180(lon::Number)

Convert longitudes measured from -180° to 180° into 0° to 360° scale.

ModelWeights.lon360to180 — Method

lon360to180(lon::Number)

Convert longitudes measured from 0° to 360° into -180° to 180° scale.

ModelWeights.longitude2EastWest — Method

longitude2EastWest(lon::Number)

Convert longitudes from -180° to 180° into 0° to 180° East/West.

ModelWeights.makeEqualWeights — Method

makeEqualWeights(data::YAXArray; use_members::Bool=true)

Create a weight vector, with equal weight for each MODEL. Distribute weight across model members if dimension=:member and usemembers is true. If usemember is false, each model member is considered as standalone model and all receive the same weight.

ModelWeights.makeSubplots — Method

makeSubplots

Arguments:

ModelWeights.mergeLoadedData — Method

mergeLoadedData(
    data_vec::Vector{YAXArray}, 
    meta_dict::Dict{String, Any}, 
    is_model_data::Bool
)

All data is assumed to be defined on the same grid.

ModelWeights.mergeMetaDataPaths — Method

mergeMetaDataPaths(meta1::Dict{String, Any}, meta2::Dict{String, Any})

Return vector of unique paths of mergee set of paths from meta1 and meta2.

ModelWeights.normalizeWeightsVariables — Method

normalizeWeightsVariables(weights_dict::Dict{String, Number})

Normalize weights for combinations of variable and diagnostic such that they sum up to 1.

Arguments:

weights_dict: maps from VARIABLEDIAGNOSTIC (e.g. tasCLIM) to weights

Return:

DimArray with dimensions 'variable' and 'diagnostic' containig the normalized weights.

ModelWeights.performanceParts — Method

performanceParts(generalizedDistances::AbstractArray, sigmaD::Number)

Compute the performance part (numerator) of the overall weight for each model.

Arguments:

generalizedDistances: contains generalized distances Di for each model

sigmaD: free model parameter for impact of performance weights

ModelWeights.plotAMOC — Method

plotAMOC(data::DimArray)

Plot AMOC strength for variable "amoc".

ModelWeights.plotDistances — Method

plotDistances(dists::AbstractArray, title::String; is_bar_plot::Bool=true)

Plot figure of distances for every combination of variable and diagnostic in dists.

Arguments:

dists: must have dimensions :variable, :diagnostic

title:

is_bar_plot:

ModelWeights.plotDistancesIndependence — Method

plotDistancesIndependence(distances::AbstractArray, dimname::String)

Arguments:

distances:

dimname:

ModelWeights.plotEnsembleSpread — Method

plotEnsembleSpread(data::AbstractArray, lon::Number, lat::Number)

Create figure with boxplots for each model in data that have several ensemble members.

ModelWeights.plotHistAtPos — Method

plotHistAtPos(data::AbstractArray, location::Dict)

Plot histogram of all data for a specific location.

Arguments:

data: dimensions 'lon' (from -180° to 180°), 'lat' (-90° to 90°)

location: must have keys 'name', 'lon', 'lat'

ModelWeights.plotTempGraph — Method

plotTempGraph

# TODO: change quantileLabels and uncertaintyRanges, the latter should contain 
# the quantile labels

ModelWeights.plotTimeseries! — Method

plotTimeseries!(ax::Axis, vals::AbstractArray;)

Plot timeseries of data vector vals.

ModelWeights.plotValsOnMap! — Method

plotValsOnMap!(fig::Figure, means::AbstractArray, title::String; colors=nothing, colorrange=nothing, highclip=(1,0,0), lowclip=(0,0,1), pos=(x=1, y=1), poslegend=nothing )

Plot contours of world with an overlayed heatmap of the input data.

Arguments:

nb_ticks: if nothing (default), just min lat/lon, max lat/lon and 0 are labeled.

pos::NamedTuple(x::Int,y::Int): position of plot in fig

ModelWeights.plotWeightContributions — Method

plotWeightContributions(independence::DimArray, performance::DimArray)

Plot performance against independence weights.

Arguments:

independence: (dims:model) normalized independence weights

performance: (dims:model) normalized performance weights

ModelWeights.plotWeights — Method

plotWeights(weights::Weights; title::String="")

Plot all weights together, i.e. overall weights, performance-only and independence-only weights.

Arguments:

weights:

title:

ModelWeights.putAtModel! — Method

putAtModel!(data::AbstractArray, dimension::Symbol, model::String, input)

ModelWeights.readDataFromDisk — Method

readDataFromDisk(target_path::String; variable::String="")

Load data from target_path. If target_path ends with '.jld2', variable must be specified, otherwise data is assumed to be binary.

ModelWeights.saveWeightsAsNCFile — Method

saveWeightsAsNCFile(weights::Weights; target_path::String)

Arguments:

weights: Weights object to be saved.

target_path: Path to where weights shall be stored.

ModelWeights.searchModelInPaths — Method

searchModelInPaths(model::String, paths::Vector{String})

Arguments:

model_id::String: has form modelname#memberID[_grid]

ModelWeights.setLookupsFromMemberToModel — Method

setLookupsFromMemberToModel(data::YAXArray, dim_names::Vector{String})

Change the lookup values for the dimension 'member' to refer to the models, i.e. they are not unique anymore. This is done in preparation to group the data by the different models.

Arguments:

data::YAXArray: has at least dimensions in dim_names.

dim_names::Vector{String}: names of dimensions to be changed, e.g. 'member',

'member1' (would be changed to 'model', 'model1').

ModelWeights.setRefPeriodInWeightsMetadata! — Method

setRefPeriodInWeightsMetadata!(meta::Dict, alias::String, timerange::String)

ModelWeights.setToSummarizedMembers! — Method

setToSummarizedMembers!(data::DataMap)

Set values for every dataset in data to the average across all members of each model.

ModelWeights.sortLongitudesWest2East — Method

sortLongitudesWest2East(data::AbstractArray)

Arrange 'data' such that western latitudes come first, then eastern latitudes.

ModelWeights.subsetModelData — Method

subsetModelData(datamap::DataMap; level::LEVEL=MEMBER)

For those datasets in datamap that specify data on the level level (i.e. have dimension :member or :model), return a new DataMap with subset of data s.t. the new datasets all have the same models (level=MODEL) or members (level=MEMBER).

If no models are shared across datasets, return the input datamap.

ModelWeights.subsetModelData — Method

subsetModelData(data::AbstractArray, shared_models::Vector{String})

Return a YAXArray that is the subset of data containing only data from the models specified in shared_models.

Takes care of metadata.

Arguments:

data: must have dimension 'member' or 'model'

shared_models: vector of models, which can either be on level of models

or members of models ('modelname#memberID[_grid]').

ModelWeights.summarizeEnsembleMembersVector — Method

summarizeEnsembleMembersVector(
    data::YAXArray, updateMeta::Bool; fn::Function=Statistics.mean

)

For each model and variable (if several given), compute a summary statistic (default: mean) across all members of that model. Instead of 'member', the returned YAXArray has dimension 'model'.

Arguments:

data::YAXArray: a YAXArray with at least dimension 'member'

updateMeta::Bool: set true if the vectors in the metadata refer to

different models. Set to false if vectors refer to different variables.

ModelWeights.throwErrorIfModelDimMissing — Method

throwErrorIfModelDimMissing(data::AbstractArray)

ModelWeights.updateGroupedDataMetadata — Method

updateGroupedDataMetadata(meta::Dict, grouped_data::DimensionalData.DimGroupByArray)

Summarize vectors in meta, refering to different models (members), such that each vector only contains N entries (N=number of models (i.e. without unique members)).

If the metadata for members of a model differ across members, the respective entry in the vector will be a vector itself.

ModelWeights.updateMetadata! — Method

updateMetadata!(
    meta::Dict{String, Any},
    source_names::Vector{String},
    is_model_data::Bool
)

Update metadata meta s.t. attributes that were only present in some files/models are set to missing.

Further key-value pairs are added concerning the data sources:

'member_names' (for model data only): vector that contains for every model a

vector with the unique names of that model's members identifier consisting of variantlabel, modelname and for CMIP6 models also grid_label.

'model_names': vector whose length is the sum of the number of all models'

members (or vector of filenames of observational data); it contains the model names for each unique model member, i.e. this vector will not be unique if any model had several members.

Arguments:

source_names::Vector{String}: model names (or filenames of observational

data)as retrieved from the original files the data was loaded from.

ModelWeights.writeDataToDisk — Method

writeDataToDisk(data, target_path::String)

Save data as Julia obj if target_path has ending '.jld2', otherwise save as binary.

Index

ModelWeights.addAnomalies!

ModelWeights.addAnomaliesGM!

ModelWeights.addDiagnostic!

ModelWeights.addGlobalMeans!

ModelWeights.addLinearTrend!

ModelWeights.addTempSTD!

ModelWeights.alignPhysics

ModelWeights.applyDataConstraints!

ModelWeights.applyModelConstraints

ModelWeights.applyWeights

ModelWeights.approxAreaWeights

ModelWeights.areaWeightedRMSE

ModelWeights.averageEnsembleMembersMatrix

ModelWeights.buildCMIP5EnsembleMember

ModelWeights.buildPathsForMetaAttrib

ModelWeights.buildPathsToDataFiles

ModelWeights.computeDistancesAllDiagnostics

ModelWeights.computeGeneralizedDistances

ModelWeights.computeGlobalMeans

ModelWeights.computeInterpolatedWeightedQuantiles

ModelWeights.computeLinearTrend

ModelWeights.computeModelDataRMSE

ModelWeights.computeModelModelRMSE

ModelWeights.computeTempSTD

ModelWeights.computeWeightedAvg

ModelWeights.computeWeights

ModelWeights.convertKgsToSv!

ModelWeights.distributeWeightsAcrossMembers

ModelWeights.filterModels

ModelWeights.filterPaths

ModelWeights.filterPathsSharedModels!

ModelWeights.filterTimeseries

ModelWeights.getAliasAsTimerange

ModelWeights.getAllCombinations

ModelWeights.getAtModel

ModelWeights.getByIdxModel

ModelWeights.getCMIPModelsKey

ModelWeights.getClosestGridPoint

ModelWeights.getDimsModel

ModelWeights.getMemberIDsFromPaths

ModelWeights.getMetaAttributesFromESMValToolConfigs

ModelWeights.getMetaDataFromYAML

ModelWeights.getModelDataDist

ModelWeights.getModelDistances

ModelWeights.getModelLogLikelihoods

ModelWeights.getPathsToData

ModelWeights.getRefPeriodAsTimerangeAndAlias

ModelWeights.getSharedModelsFromPaths

ModelWeights.getTimerangeAsAlias

ModelWeights.getUncertaintyRanges

ModelWeights.getUniqueMemberIds

ModelWeights.get_optional_fields_config

ModelWeights.independenceParts

ModelWeights.isValidDataAndWeightInput

ModelWeights.joinDataMaps

ModelWeights.kelvinToCelsius

ModelWeights.kelvinToCelsius!

ModelWeights.latitude2NorthSouth

ModelWeights.loadDataFromESMValToolRecipes

ModelWeights.loadDataFromMetadata

ModelWeights.loadDataFromYAML

ModelWeights.loadPreprocData

ModelWeights.loadWeightsAsDimArray

ModelWeights.lon180to360

ModelWeights.lon360to180

ModelWeights.longitude2EastWest

ModelWeights.makeEqualWeights

ModelWeights.makeSubplots

ModelWeights.mergeLoadedData

ModelWeights.mergeMetaDataPaths

ModelWeights.normalizeWeightsVariables

ModelWeights.performanceParts

ModelWeights.plotAMOC

ModelWeights.plotDistances

ModelWeights.plotDistancesIndependence

ModelWeights.plotEnsembleSpread

ModelWeights.plotHistAtPos

ModelWeights.plotTempGraph

ModelWeights.plotTimeseries!

ModelWeights.plotValsOnMap!

ModelWeights.plotWeightContributions

ModelWeights.plotWeights

ModelWeights.putAtModel!

ModelWeights.readDataFromDisk

ModelWeights.saveWeightsAsNCFile

ModelWeights.searchModelInPaths

ModelWeights.setLookupsFromMemberToModel

ModelWeights.setRefPeriodInWeightsMetadata!

ModelWeights.setToSummarizedMembers!

ModelWeights.sortLongitudesWest2East

ModelWeights.subsetModelData

ModelWeights.subsetModelData

ModelWeights.summarizeEnsembleMembersVector

ModelWeights.throwErrorIfModelDimMissing

ModelWeights.updateGroupedDataMetadata

ModelWeights.updateMetadata!

ModelWeights.writeDataToDisk