Public API

Index

ModelWeights.Weights.MMEWeights

ModelWeights.Data.add

ModelWeights.Data.alignPhysics

ModelWeights.Data.anomalies

ModelWeights.Data.apply

ModelWeights.Data.areaWeightMatrix

ModelWeights.Data.areaWeightedRMSE

ModelWeights.Data.average

ModelWeights.Data.checkConstraint

ModelWeights.Data.combineAll

ModelWeights.Data.combineModelsFromMultipleFiles

ModelWeights.Data.constrainMetaData!

ModelWeights.Data.convertToYAX

ModelWeights.Data.defineDataMap

ModelWeights.Data.defineDataMap

ModelWeights.Data.defineDataMap

ModelWeights.Data.dict2YAX

ModelWeights.Data.dimNames

ModelWeights.Data.distancesData

ModelWeights.Data.distancesData

ModelWeights.Data.distancesData

ModelWeights.Data.distancesModels

ModelWeights.Data.distancesModels

ModelWeights.Data.filterPathsSharedModels

ModelWeights.Data.filterPathsSharedModels

ModelWeights.Data.generalizedDistances

ModelWeights.Data.generalizedDistances

ModelWeights.Data.getAtModel

ModelWeights.Data.getByIdxModel

ModelWeights.Data.getCMIPModelsKey

ModelWeights.Data.globalMeans

ModelWeights.Data.indexModel

ModelWeights.Data.indicesTimeseries

ModelWeights.Data.individuatePath

ModelWeights.Data.joinDicts

ModelWeights.Data.kelvinToCelsius

ModelWeights.Data.loadDataFromESMValToolRecipes

ModelWeights.Data.loadDataFromYAML

ModelWeights.Data.loadDataMapCore

ModelWeights.Data.loadModelsFromCSV

ModelWeights.Data.loadPreprocData

ModelWeights.Data.lon180to360

ModelWeights.Data.lon360to180

ModelWeights.Data.longitudesEastWest

ModelWeights.Data.memberIDFromFilenameMeta

ModelWeights.Data.memberIDsFromPaths

ModelWeights.Data.mergeMetaDataFromMultipleFiles

ModelWeights.Data.metaDataChecksCMIP

ModelWeights.Data.metaDataFromESMValToolRecipes

ModelWeights.Data.metaDataFromYAML

ModelWeights.Data.modelDim

ModelWeights.Data.modelDims

ModelWeights.Data.nbModelMembers

ModelWeights.Data.normalizeDict

ModelWeights.Data.parseFilename

ModelWeights.Data.putAtModel!

ModelWeights.Data.quantile

ModelWeights.Data.readDataFromDisk

ModelWeights.Data.resolvePathsFromMetaData

ModelWeights.Data.setDim

ModelWeights.Data.setLookupsFromMemberToModel

ModelWeights.Data.sharedLevelMembers

ModelWeights.Data.sharedLevelModels

ModelWeights.Data.sharedModels

ModelWeights.Data.sortLongitudesEast2West

ModelWeights.Data.sortLongitudesWest2East

ModelWeights.Data.subsetDataMap

ModelWeights.Data.subsetMeta

ModelWeights.Data.subsetMeta!

ModelWeights.Data.subsetModelData

ModelWeights.Data.subsetModelData

ModelWeights.Data.summarizeMembers

ModelWeights.Data.summarizeMembers

ModelWeights.Data.summarizeMembers!

ModelWeights.Data.summarizeMembersMatrix

ModelWeights.Data.summarizeMembersVector

ModelWeights.Data.throwErrorIfDimMissing

ModelWeights.Data.throwErrorIfDimMissing

ModelWeights.Data.uncertaintyRanges

ModelWeights.Data.updateGroupedDataMetadata

ModelWeights.Data.writeDataToDisk

ModelWeights.Plots.convertKgsToSv!

ModelWeights.Plots.getClosestGridPoint

ModelWeights.Plots.latitude2NorthSouth

ModelWeights.Plots.longitude2EastWest

ModelWeights.Plots.makeSubplots

ModelWeights.Plots.plotAMOC

ModelWeights.Plots.plotDistances

ModelWeights.Plots.plotDistancesIndependence

ModelWeights.Plots.plotEnsembleSpread

ModelWeights.Plots.plotHistAtPos

ModelWeights.Plots.plotTempGraph

ModelWeights.Plots.plotTimeseries!

ModelWeights.Plots.plotValsOnMap!

ModelWeights.Plots.plotWeights

ModelWeights.Timeseries.computeLinearTrend

ModelWeights.Timeseries.filterTimeseries

ModelWeights.Timeseries.linearTrend

ModelWeights.Weights.applyWeights

ModelWeights.Weights.climwipWeights

ModelWeights.Weights.climwipWeights

ModelWeights.Weights.crps

ModelWeights.Weights.crps

ModelWeights.Weights.crpss

ModelWeights.Weights.crpss_weighted

ModelWeights.Weights.crpss_weighted

ModelWeights.Weights.distributeWeightsAcrossMembers

ModelWeights.Weights.equalWeights

ModelWeights.Weights.independenceWeights

ModelWeights.Weights.performanceWeights

ModelWeights.Weights.rps

ModelWeights.Weights.saveWeightsAsNCFile

ModelWeights.Weights.weightSamples

ModelWeights.Weights.weightedAvg

Functions

ModelWeights.Data.add — Method

add(data::YAXArray, dimension::Symbol)

Wrapper function for sum that removes 'dimension' from the returned YAXArray.

ModelWeights.Data.alignPhysics — Method

alignPhysics(data::YAXArray, members::Vector{String})

Return new YAXArray with the models retained from data that share the same physics as the respective model's members in members. All other models that are in data but for which no member is specified in members are also retained.

ModelWeights.Data.anomalies — Method

anomalies(orig_data::YAXArray, ref_data::YAXArray)

Compute difference of orig_data and ref_data.

ModelWeights.Data.apply — Method

apply(
    dm::DataMap,
    fn::Function,
    args...; 
    ids::AbstractVector{T} = Vector{String}(),
    ids_new::AbstractVector{T} = Vector{String}(),
    kwargs...
)

Apply fn with positional arguments args and keyword arguments kwargs and return result as a new DataMap where keys are ids_new if provided, otherwise same keys as in dm are used.

Arguments:

dm::DataMap: data.

fn::Function: function to be applied.

args...: positional arguments for fn.

ids::AbstractVector{T}=Vector{String}(): keys for data on which fn is applied; if empty all keys of dm are used.

ids_new::AbstractVector{T}=Vector{String}(): keys used in new datamap; if empty ids are used instead.

kwargs...: keyword arguments for fn.

ModelWeights.Data.areaWeightMatrix — Method

areaWeightMatrix(
    latitudes::AbstractVector{<:Number}, mask::YAXArray{T}
) where {T <: Union{Missing, Bool}}

Return matrix of size length(longitudes) x length(latitudes) with normalized area weights (approximated based on latitudes) and set to 0 at positions where mask is true.

Arguments:

latitudes::AbstractVector{<:Number}: to approximate area weights with cosine of latitudes.

mask::AbstractArray{T}: lon, lat as first and second dimension respectively where lat must

have same length as latitudes.

ModelWeights.Data.areaWeightedRMSE — Method

areaWeightedRMSE(m1::AbstractArray, m2::AbstractArray, aw_mat::AbstractArray)

Compute the area weighted (approximated by cosine of latitudes in radians) root mean squared error between m1 and m2.

Arguments:

m1: must have dimensions 'lon', 'lat' as first dimensions.

m2: must have dimensions 'lon', 'lat' as first dimensions.

aw_mat: matrix with area weights, of same size as m1 and m2, that will be normalized.

ModelWeights.Data.average — Method

average(data::YAXArray, dimension::Symbol)

Wrapper function for mean that removes 'dimension' from the returned YAXArray.

ModelWeights.Data.checkConstraint — Method

checkConstraint(constraint::Union{Dict, Nothing})

Throw error if entries (except for level_shared) do not map to a vector of Strings.

Used to provide specific error messages of values that were not specified correctly.

ModelWeights.Data.combineAll — Method

combineAll(v::Vararg{Vector{String}}; sep::String="_")

Generate vector with all possible combinations of strings, each of which is a concatenation of elements from each input vector in the given order, concatenated by sep (default: _).

Arguments

v::Vararg{Vector{String}}: A variable number of input vectors.

sep::String: seperator

Example

julia> ModelWeights.Data.combineAll(["tos", "tas"], ["CLIM"])
2-element Vector{String}:
 "tos_CLIM"
 "tas_CLIM"

ModelWeights.Data.combineModelsFromMultipleFiles — Method

combineModelsFromMultipleFiles(
    data::Vector{AbstractArray}; 
    model_names::Vector{String} = Vector{String}(),
    meta::Union{Dict{String, T}, Nothing} = nothing,
    dtype::String = "undef",
    new_dim::Symbol = :model,
    sorted::Bool = true

) where T <: Any

Combine data from different files into a single YAXArray. The meta data of the returned YAXArray is a combination of all values from all datasets using vectors with missing values for datasets that didn't have the respective metadata entry in their metadata, further combined with the properties in meta if provided.

All elements in data must share all dimensions except for time (if present). For timeseries data, the time dimension may cover different ranges. In that case the maximal overlapping time period is used.

The combined YAXArray has the additional dimension new_dim (default: :model) with names as values. If names is not provided, default values 'model1', 'model2, etc. are used instead.

If sorted is true, model dimension of returned data is sorted alphabetically and the vector entries in the metadata dictionary of the returned array are also sorted accordingly.

ModelWeights.Data.constrainMetaData! — Method

constrainMetaData!(meta_attributes::Vector{MetaData}, constraint::Dict)

Subset entries in meta_attributes so that only those with properties specified in constraint remain.

Arguments

constraint::Dict: Mapping to vector specifiying the properties of which at least one

must be present for an id to be retained.

ModelWeights.Data.convertToYAX — Method

convertToYAX(dm::DataMap; dim_name::Symbol = :diagnostic)

Convert a DataMap into a YAXArray. All entries in dm must have the same dimensions.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    path_data_dir::String, 
    id::String;
    meta_data::Dict{String, T} = Dict{String, Any}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "cmip",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entry id with the data (model and/or obs depending on dtype) from all .nc files in paths and all .nc files in all directories in paths, possibly constraint by constraint.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    paths_data_dirs::Vector{String}, 
    data_ids::Vector{String};
    meta_data::Vector{Dict{String, T}} = Vector{Dict{String, Any}}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "cmip",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entries data_ids with the data (model and/or obs depending on dtype) from all .nc files in all directories in paths_data_dirs, possibly constraint by constraint.

ModelWeights.Data.defineDataMap — Method

defineDataMap(
    paths_data_dirs::Vector{String}, 
    data_ids::Vector{String};
    meta_data::Vector{Dict{String, T}} = Vector{Dict{String, Any}}(),
    constraint::Union{Dict, Nothing} = nothing,
    sorted::Bool = true, 
    dtype::String = "cmip",
    filename_format::Union{Symbol, String} = :cmip
) where T <: Any

Return DataMap with entries data_ids with the data (model and/or obs depending on dtype) from all .nc files in all directories in paths_dirs, possibly constraint by constraint.

For every loaded dataset (entry in built DataMap), files are loaded from several directories; each entry in paths_data_dirs points the the vector of data directories from where data is loaded for that dataset.

ModelWeights.Data.dict2YAX — Method

dict2YAX(data::Dict{String, <:Number})

Convert dictionary data into a YAXArray with new dimension dim_name (default is :diagnostic) whose lookup names are the keys of data.

ModelWeights.Data.dimNames — Method

dimNames(data::YAXArray)

Return the names of the dimensions of data as vector of symbols.

ModelWeights.Data.distancesData — Method

distancesData(
    models::AbstractArray, observations::AbstractArray, latitudes::AbstractVector
)

Compute the distance as the area-weighted RMSE between model predictions and observations.

Neither the observations nor the model data must contain missing values.

Arguments:

models: with dimensions 'lon', 'lat', 'model'/'member' (in this order)

observations: with dimensions 'lon', 'lat' (in this order)

latitudes: for computing area weight matrices

ModelWeights.Data.distancesData — Method

distancesData(model_data::DataMap, obs_data::DataMap, diagnostics::Vector{String})

Compute RMSEs between models and observations for diagnostics.

Arguments:

diagnostics::Vector{String}: keys for which values must be provided in model_data and

obs_data.

ModelWeights.Data.distancesData — Method

distancesData(models::YAXArray, observations::YAXArray)

Compute the distance as the area-weighted RMSE between model predictions and observations.

If several observational datasets are present, the average across all is taken.

Arguments:

models::YAXArray: must have dimensions 'lon', 'lat' and one of 'model', 'member'.

observations::YAXArray: may have dimension 'model', then average across datasets is used.

ModelWeights.Data.distancesModels — Method

distancesModels(model_data::DataMap, diagnostics::Vector{String})

Compute the model-model distances for model_data at every diagnostic in diagnostics and return a vector of YAXArrays.

ModelWeights.Data.distancesModels — Method

distancesModels(data::YAXArray)

Compute the area weighted RMSE between model predictions for each pair of models.

Arguments:

data::YAXArray: first two dimensions must be 'lon', 'lat', third 'model' or 'member'.

ModelWeights.Data.filterPathsSharedModels — Method

filterPathsSharedModels(
    paths::Vector{String}, 
    shared_models::Vector{String}, 
    fn_format::Union{Symbol, String}
)

Every vector of paths in all_paths is filtered s.t. it only contains models or model members given in shared_models.

Arguments:

paths: contains paths to data files

ModelWeights.Data.filterPathsSharedModels — Method

filterPathsSharedModels(
    all_paths::Vector{Vector{String}}, 
    level_shared::Level, 
    fn_format::Union{Symbol, String}
)

Arguments:

all_paths: every entry refers to the paths to data files for the respective dataset

ModelWeights.Data.generalizedDistances — Method

generalizedDistances(
    distances_all::Vector{<:YAXArray}, diagnostics::Vector{String}, weights::YAXArray
)

Arguments:

distances_all::Vector{<:YAXArray}: each entry refers to the distances of one diagnostic,

each must have dimension 'model', 'member' or 'member1' and 'member2' (must be identical for every entry!).

diagnostics:: Vector{String}: names of diagnostics.

weights::YAXArray: weights assigned to each diagnostic.

norm_avg_members::Bool: if true (default), average distances per model are computed

BEFORE computing the median that is used to normalize the distances (seperately for each diagnostic); if false, median is computed across all individual members of each diagnostic.

ModelWeights.Data.generalizedDistances — Method

generalizedDistances(
    distances_all::YAXArray, weights::YAXArray; norm_avg_members::Bool = true
)

For every diagnostic in distances_all, compute the weighted sum of all diagnostics.

Arguments:

distances_all::YAXArray: must have dimension :diagnostic.

weights::YAXArray: weights assigned to each diagnostic.

norm_avg_members::Bool: if true (default), average distances per model are computed

BEFORE computing the median that is used to normalize the distances (seperately for each diagnostic); if false, median is computed across all individual members of each diagnostic.

ModelWeights.Data.getAtModel — Method

getAtModel(data::YAXArray, dimension::Symbol, model::String)

Return data where dimension (member or model) has value model.

ModelWeights.Data.getByIdxModel — Method

getByIdxModel(data::YAXArray, dimension::Symbol, indices::Vector)

Return data where dimension (member or model) has value model.

ModelWeights.Data.getCMIPModelsKey — Method

getCMIPModelsKey(meta::Dict)

Return the respective key to retrieve model names in CMIP6 ('sourceid') and CMIP5 ('modelid') data.

If both keys are present, 'source_id' used in CMIP6 models is returned, if none is present, throw ArgumentError.

ModelWeights.Data.globalMeans — Method

globalMeans(data::YAXArray)

Return a YAXArray with area-weighted global means for each model in data.

Missing data is accounted for in the area-weights.

Arguments:

data::YAXArray: must have dimensions 'lon' and 'lat'.

ModelWeights.Data.indexModel — Method

indexModel(data::YAXArray, model_dims::Tuple{Symbol}, indices::Vector{Int})

Return data at model dimensions model_dims at indices.

ModelWeights.Data.indicesTimeseries — Method

indicesTimeseries(times::Vector, constraint_ts::Dict)

Get indices of times that are exactly aligned with start_y and end_y given in constraint_ts. If no values are given, return indices for entire vector times.

ModelWeights.Data.individuatePath — Method

individuatePath(target::String)

If file at target path exists, concatenate target with current timestep, otherwise simply return target.

ModelWeights.Data.joinDicts — Method

joinDicts(v::Dict...; warn_msg::String="")

ModelWeights.Data.kelvinToCelsius — Method

kelvinToCelsius(data::YAXArray)

Return a copy of data with values given in Kelvin covnerted into Degree Celsius.

ModelWeights.Data.loadDataFromESMValToolRecipes — Method

loadDataFromESMValToolRecipes(
    path_data::String,
    path_recipes::String;
    dir_per_var::Bool = true,
    constraint::Union{Dict, Nothing} = nothing,
    preview::Bool = false,
    sorted::Bool = true
)

Load the data from the ESMValTool recipes at path_recipes or, if preview is true, load meta data only.

Arguments:

path_data: top level directory were data is stored.

path_recipes: directory were ESMValTool recipes are stored; these must be the versions

in the run folder generated by ESMValTool named RECIPENAME_filled.yml.

dir_per_var: set to true (default) if there is a subdirectory in path_data for every

climate variable to be loaded.

constraint: TODO!

preview: set to true if only meta data should be loaded (default: false).

sorted: if true (default), the data is sorted alphabetically wrt model names.

ModelWeights.Data.loadDataFromYAML — Method

loadDataFromYAML(
    content::Dict;
    constraint::Union{Dict, Nothing} = nothing,
    preview::Bool = false,
    sorted::Bool = true,
    dtype::String = "cmip"
)

Return a DataMap-instance that contains the data specified in content, potentially constraint by values in constraint.

Arguments:

preview::Bool: if true (default: false), return metadata and corresponding paths without

actually loading any data.

sorted::Bool: if true (default), model dimension is sorted alphabetically.

dtype::String: if set to "cmip", model dimension of returned data have model names as values.

ModelWeights.Data.loadDataMapCore — Method

loadDataMapCore(
    all_paths::Vector{Vector{String}},
    ids::Vector{String},
    constraints::Vector{Dict};
    dtype::String = "cmip",
    filename_format::Union{Symbol, String} = :cmip,
    sorted::Bool = true,
    preview::Bool = false,
    meta_data::Union{Vector{Dict{String, T}}, Nothing} = nothing
) where T <: Any

Load a DataMap instance with keys ids that map to data at all_paths, where every subvector refers the the paths of the data of a single dataset.

ModelWeights.Data.loadModelsFromCSV — Method

loadModelsFromCSV(
    path::String, col_models::String; col_variants::Union{String,Nothing}=nothing
)

Return a vector with models retrieved from csv file at path. If col_variants is provided, returned models are on level of model members (MODEL#variant, e.g. AWI-ESM#r1i1p1f1).

ModelWeights.Data.loadPreprocData — Method

loadPreprocData(
    paths::Vector{String},
    filename_format::Union{Symbol, String};
    sorted::Bool = true, 
    dtype::String = "cmip",
    names::Vector{String} = Vector{String}(),
    meta_info::Union{Dict{String, T}, Nothing} = nothing,
    constraint_ts::Union{Dict, Nothing} = nothing
) where T <: Any

Return data loaded from paths as single YAXArray.

Each path points to a different model, i.e. the data for one dataset is loaded from multiple files and all datasets in paths must share the same dimensions. Which variable is loaded is inferred from the filenames (in paths), or if meta_info has key 'variable', the respective value is used.

ModelWeights.Data.lon180to360 — Method

lon180to360(lon::Number)

Convert longitudes measured from -180° to 180° into 0° to 360° scale. For western hemisphere (negative longitudes) add 360.

ModelWeights.Data.lon360to180 — Method

lon360to180(lon::Number)

Convert longitudes measured from 0° to 360° into -180° to 180° scale.

ModelWeights.Data.longitudesEastWest — Method

sortLongitudesWest2East(data::AbstractArray)

Arrange 'data' such that western latitudes come first, then eastern latitudes.

Return NamedTuple with fields 'east' and 'west' pointing to vectors of indices for sorted longitudes.

ModelWeights.Data.memberIDFromFilenameMeta — Method

memberIDFromFilenameMeta(fn_meta::FilenameMeta)

Return a string of the form modelname#memberID[_grid] that identifies the corresponding model member.

For CMIP6 models the abbreviation of the grid is added to the model name.

ModelWeights.Data.memberIDsFromPaths — Method

memberIDsFromPaths(all_paths::Vector{String})

For every path in all_paths return a string of the form modelname#memberID[grid] that identifies the corresponding model member. Filenames must follow the CMIP-standard (projnamemipexpidvariable[_grid].nc).

For CMIP6 models the abbreviation of the grid is added to the model name.

ModelWeights.Data.mergeMetaDataFromMultipleFiles — Method

mergeMetaDataFromMultipleFiles(data::Vector{<:YAXArray})

Combine arrays in data into a single YAXArray with meta combined from all datasets into lists, with missing if key wasnt in a dataset.

ModelWeights.Data.metaDataChecksCMIP — Method

metaDataChecksCMIP(meta::Dict, path::String)

Check model names as retrieved from the metadata for potential inconsistencies wrt filename.

ModelWeights.Data.metaDataFromESMValToolRecipes — Method

metaDataFromESMValToolRecipes(
    base_path_configs::String; constraint::Union{Dict, Nothing} = nothing
)

Read variable, statistic, experiment and timerange/alias values from ESMValTool recipes stored at base_path_configs into a vector of Dictionaries storing the respective readoff values.

ModelWeights.Data.metaDataFromYAML — Method

metaDataFromYAML(ds::Dict)

Return metadata of data specified in ds.

Arguments:

ds: content of config yaml file specifying meta attributes and paths of data.

ModelWeights.Data.modelDim — Method

modelDim(data::YAXArray)

Return the model dimension of data which must either be :model or :member.

If none is present, throw ArgumentError.

ModelWeights.Data.modelDims — Method

modelDims(data::YAXArray)

Return vector of dimensions of data that contain either 'model' or 'member'.

ModelWeights.Data.nbModelMembers — Method

nbModelMembers(data::YAXArray)

Return a dictionary mapping from models in data to the number of its members.

ModelWeights.Data.normalizeDict — Method

normalizeDict(data::Dict{String, <:Number})

Normalize values for every entry in data such that they sum up to 1. If remove_zero is true (default), the returned dictionary does not contain entries for which values were 0.

ModelWeights.Data.parseFilename — Method

parseFilename(filename::String, format::String)

Retrieve information from filename and returns it as an object of type FilenameMeta.

ModelWeights.Data.putAtModel! — Method

putAtModel!(data::YAXArray, dimension::Symbol, model::String, input)

ModelWeights.Data.quantile — Method

quantile(samples::Vector{<:Number}, p::Number)

Compute the p-th quantile.

ModelWeights.Data.readDataFromDisk — Method

readDataFromDisk(target_path::String; variable::String="")

Load data from target_path. If target_path ends with '.jld2', variable must be specified, otherwise data is assumed to be binary.

ModelWeights.Data.resolvePathsFromMetaData — Method

resolvePathsFromMetaData( 
    meta::MetaData,
    base_path::String,
    dir_per_var::Bool;
    constraint::Union{Dict, Nothing} = nothing
)

Return paths to data files for data specified in meta, possibly constraint by values in constraint. The paths were the data is stored is expected to follow the following structure (corresponding to the output from ESMValTool used for preprocessing the data):

base_path is the top-level directory. If dir_per_var is true, base_path is assumed to have a (or several) subdirectory for each climate variable with VAR as part of the subdirectory's name (e.g. _tas, cmip5tas, etc.). These subdirectories may be constraint by containing at least one of the values in constraint["base_subdirs"].

Let BASE refer to base_path, or respectively, to the subdirectories for the climate variables. Then the following structure is: BASE/preproc/meta.alias/meta.subdir. In ESMValTool, meta.alias corresponds to the (self-chosen) name under the section 'diagnostics' and meta.subdir to the (self-chosen) name under the section 'variables'.

The returned paths are the paths to all files within this directory, possibly constraint by the filenames containig at least one string in constraint["projects"] and respectively at least one string in constraint["models"].

ModelWeights.Data.setDim — Method

setDim(
    data::YAXArray, 
    dim::Union{String, Symbol},
    dim_name::Union{Nothing, String, Symbol},
    dim_vals::Union{Nothing, Vector{String}}
)

Rename dimension dim to dim_name and/or its values to dim_vals. If dim_name or dim_vals is nothing, name/vals don't change.

Arguments:

data::YAXArray:

dim::Union{String, Symbol}: Name of the dimension to be changed.

dim_name::Union{Nothing, String, Symbol}: New dimension name.

dim_vals::Union{Nothing, Vector{String}}: New dimension values.

ModelWeights.Data.setLookupsFromMemberToModel — Method

setLookupsFromMemberToModel(data::YAXArray, dim_names::Vector{String})

Change the lookup values for the dimension 'member' to refer to the models, i.e. they are not unique anymore. This is done in preparation to group the data by the different models.

Arguments:

data::YAXArray: has at least dimensions in dim_names.

dim_names::Vector{String}: names of dimensions to be changed, e.g. 'member',

'member1' (would be changed to 'model', 'model1').

ModelWeights.Data.sharedLevelMembers — Method

sharedLevelMembers(data::DataMap)

Return vector of model members that are shared across all entries in data.

ModelWeights.Data.sharedLevelModels — Method

sharedLevelModels(data::DataMap)

Return vector of models (on level of models, not members) that are shared across all entries in data.

ModelWeights.Data.sharedModels — Method

sharedModels(
    all_paths::Vector{Vector{String}},
    level_shared::Level,
    fn_format::Union{Symbol, String}
)

Arguments:

all_paths: every entry refers to the paths to data files for the respective dataset

ModelWeights.Data.sortLongitudesEast2West — Method

sortLongitudesWest2East(data::AbstractArray)

Arrange 'data' such that western latitudes come first, then eastern latitudes.

ModelWeights.Data.sortLongitudesWest2East — Method

sortLongitudesWest2East(data::AbstractArray)

Arrange 'data' such that western latitudes come first, then eastern latitudes.

ModelWeights.Data.subsetDataMap — Method

subsetDataMap(data::DataMap, ids::Vector{String})

Return new DataMap with data from data at ids.

ModelWeights.Data.subsetMeta! — Method

subsetMeta(meta::Dict, indices::Vector{<:Int}; simplify::Bool = false)

If simplify is true, use single value when all remaining elements in a vector in meta are identical, otherwise just keep vector.

Arguments:

meta:

indices: indices to remain in vectors mapped to in meta.

ModelWeights.Data.subsetMeta — Method

subsetMeta(meta::Dict, indices::Vector{<:Int}; simplify::Bool = false)

If simplify is true, use single value when all remaining elements in a vector in meta are identical, otherwise just keep vector.

Arguments:

meta:

indices: indices to remain in vectors mapped to in meta.

ModelWeights.Data.subsetModelData — Function

subsetModelData(dm::DataMap, level::Level=MEMBER_LEVEL; ids::Vector{String}=Vector{String}())

For those datasets in datamap that specify data on the level level (i.e. have dimension :member or :model), return a new DataMap with subset of data s.t. the new datasets all have the same models (MODELLEVEL) or members (MEMBERLEVEL).

If no models are shared across datasets, return the input datamap.

ModelWeights.Data.subsetModelData — Method

subsetModelData(data::YAXArray, shared_models::Vector{String})

Return subset of data containing only data from models in shared_models.

Takes care of metadata.

Arguments:

data: must have a dimension that contains 'member' or 'model'

shared_models: models, which can either be on level of models or members of models

('modelname#memberID[_grid]').

ModelWeights.Data.summarizeMembers! — Method

summarizeMembers!(data::DataMap; fn::Function=Statistics.mean)

Set values for every dataset in data to the average across all members of each model.

ModelWeights.Data.summarizeMembers — Method

summarizeMembers(data::DataMap; fn::Function=Statistics.mean)

Return new DataMap containing every dataset in data summarized by applying fn (default: mean) to all members of each model.

ModelWeights.Data.summarizeMembers — Method

summarizeMembers(data::YAXArray; fn::Function=Statistics.mean)

For each model compute a summary statistic (default: mean) across all its members.

The returned YAXArray has dimension 'model'.

Arguments:

data::YAXArray: YAXArray; must have dimension 'member'

updateMeta::Bool: set true if the vectors in the metadata refer to

different models. Set to false if vectors refer to different variables.

fn::Function: Function to be applied on data

ModelWeights.Data.summarizeMembersMatrix — Method

summarizeMembersMatrix(data::AbstractArray, updateMeta::Bool; fn::Function=Statistics.mean)

Compute the average across all members of each model for each given variable for model to model data, e.g. distances between model pairs.

Arguments:

data: with at least dimensions 'member1', 'member2'

updateMeta: set true if the vectors in the metadata refer to different models.

Set to false if vectors refer to different variables for instance.

ModelWeights.Data.summarizeMembersVector — Method

summarizeMembersVector(data::YAXArray; fn::Function=Statistics.mean)

For each model compute a summary statistic (default: mean) across all its members. The returned YAXArray has dimension :model (instead of :member).

Arguments:

data::YAXArray: YAXArray; must have dimension 'member' and at least one other arbitrary dimension

fn::Function: Function to be applied on data

ModelWeights.Data.throwErrorIfDimMissing — Method

throwErrorIfDimMissing(data::YAXArray)

ModelWeights.Data.throwErrorIfDimMissing — Method

throwErrorIfDimMissing(data::YAXArray, dimensions::Vector{Symbol}; include::Symbol=:all)

Throw ArgumentError if data does not have ALL dimensions in dimensions when include=:all (default), or if data does not have ANY dimension in dimensions.

ModelWeights.Data.uncertaintyRanges — Method

uncertaintyRanges(
    data::AbstractArray; w::Union{DimArray, Nothing}=nothing, quantiles=[0.167, 0.833]
)

Compute weighted quantiles of timeseries data.

Arguments:

data::AbstractArray: must have dimensions 'time', 'model'.

w::Union{DimArray, Nothing}=nothing: must have dimensions 'model' and 'weight', each

'weight'-vector must sum up to 1.

quantiles=[0.167, 0.833]: vector with two entries between 0 and 1 representing the

lower and upper bound in this order.

ModelWeights.Data.updateGroupedDataMetadata — Method

updateGroupedDataMetadata(meta::Dict, grouped_data::DimensionalData.DimGroupByArray)

Summarize vectors in meta, refering to different models (members), such that each vector only contains N entries (N=number of models (i.e. without unique members)).

If the metadata for members of a model differ across members, the respective entry in the vector will be a vector itself.

ModelWeights.Data.writeDataToDisk — Method

writeDataToDisk(data, target_path::String)

Save data as Julia obj if target_path has ending '.jld2', otherwise save as binary. If file at target_path already exists, timestamp is added if overwrite is false (default).

ModelWeights.Plots.convertKgsToSv! — Method

convertKgsToSv!(vec:YAXArray)

Convert data given in unit 'kg s-1' into Sverdrups (Sv).

ModelWeights.Plots.getClosestGridPoint — Method

getClosestGridPoint(location::Dict, longitudes::Vector, latitudes::Vector)

Find the grid point in grid defined by longitudes and latitudes that is closest to location.

Arguments

location::Dict: 'lon', 'lat' of position for which closest grid point is

returned, lon must be given from -180° to 180°

longitudes::Vector: grid longitudes measured from -180° to 180°

latitudes::Vector: grid latitudes measured from -90° to 90°

ModelWeights.Plots.latitude2NorthSouth — Method

latitude2NorthSouth(lat::Number)

Convert latitudes from -90° to 90° into 0° to 90° North/South.

ModelWeights.Plots.longitude2EastWest — Method

longitude2EastWest(lon::Number)

Convert longitudes from -180° to 180° into 0° to 180° East/West.

ModelWeights.Plots.makeSubplots — Method

makeSubplots

Arguments:

ModelWeights.Plots.plotAMOC — Method

plotAMOC(data::DimArray)

Plot AMOC strength for variable "amoc".

ModelWeights.Plots.plotDistances — Method

plotDistances(dists::AbstractArray, title::String; is_bar_plot::Bool=true)

Plot figure of distances.

Arguments:

dists::AbstractArray:

title::String:

is_bar_plot::Bool = true:

ModelWeights.Plots.plotDistancesIndependence — Method

plotDistancesIndependence(
    distances::AbstractArray, dimname::String; title::String="Generalized distances Sij"
)

Arguments:

distances::AbstractArray:

dimname::String:

title::String:

ModelWeights.Plots.plotEnsembleSpread — Method

plotEnsembleSpread(
    data::YAXArray; fn::Function=Statistics.median, title::String="", ylabel::String=""
)

Plot models against values of all model members. Add summary statistics (default: median) when computed across all model members and when average values across members of each model are computed first.

Arguments:

data::YAXArray: must have single dimension :member.

fn::Function=Statistics.median: function to summarize data.

title::String: (optional) plot title.

ylabel::String: (optional) label for y-axis.

ModelWeights.Plots.plotHistAtPos — Method

plotHistAtPos(data::AbstractArray, location::Dict)

Plot histogram of all data for a specific location.

Arguments:

data: dimensions 'lon' (from -180° to 180°), 'lat' (-90° to 90°)

location: must have keys 'name', 'lon', 'lat'

ModelWeights.Plots.plotTempGraph — Method

plotTempGraph

# TODO: change quantileLabels and uncertainty_ranges, the latter should contain 
# the quantile labels

ModelWeights.Plots.plotTimeseries! — Method

plotTimeseries!(ax::Axis, vals::AbstractArray;)

Plot timeseries of data vector vals.

ModelWeights.Plots.plotValsOnMap! — Method

plotValsOnMap!(fig::Figure, means::AbstractArray, title::String; colors=nothing, colorrange=nothing, highclip=(1,0,0), lowclip=(0,0,1), pos=(x=1, y=1), poslegend=nothing )

Plot contours of world with an overlayed heatmap of the input data.

Arguments:

pos::NamedTuple(x::Int,y::Int): position of plot in fig

color_range::Union{Nothing, Tuple}: Real values, outside this range, a single color is used.

ModelWeights.Plots.plotWeights — Method

plotWeights(weights::YAXArray; title::String="", sort_by::String="")

Plot weights sorted by weight dimension sort_by of weights if given, otherwise, sorted by first value in weight dimension.

Equal weights line is added to the plot.

Arguments:

weights::YAXArray: with dimensions :weight and :model

title::String:

sort_by::String: value of weight dimension according to which data is sorted

ModelWeights.Timeseries.computeLinearTrend — Method

computeLinearTrend(x::Vector, y::Vector; full_predictions::Bool = false)

Regress y on x and return slope (default) or predicted y for every element in x (if full_predictions=true)

Arguments:

x::Vector:

y:Vector:

full_predictions::Bool: if false, return slope, otherwise return all values predicted for x.

ModelWeights.Timeseries.filterTimeseries — Method

function filterTimeseries(
    data::YAXArray,
    start_year::Number,
    end_year::Number;
    only_models_non_missing_vals::Bool = true
)

Arguments:

data::YAXArray:

start_year::Number: set to -Inf to not filter for start time.

end_year::Number: set to Inf to not filter for end time.

only_models_non_missing_vals::Bool = true:

ModelWeights.Timeseries.linearTrend — Method

linearTrend(data::YAXArray; full_predictions::Bool=true)

Compute linear trend as ordinary least squares for timeseries data and return slope (default) or predicted y for every element in x (if full_predictions=true)

Arguments:

data::YAXArray: must have dimension :time.

full_predictions::Bool:

ModelWeights.Weights.applyWeights — Method

applyWeights(model_data::YAXArray, weights::YAXArray)

Compute the weighted average of model data data with given weights.

If weights were computed for a superset of the models in data, they are normalized and applied to the subset. Only weights per model (not members) are considered for now, in the future, members should be considered too.

Arguments:

model_data::YAXArray: model predictions. If given for model members, the predictions

of each model are considered the average value of all members of the respective model.

weights::YAXArray: if given for each member of a model, these will be summed up to

yield one value per model.

ModelWeights.Weights.climwipWeights — Method

climwipWeights(
    model_data::DataMap,
    obs_data::DataMap,
    config::ConfigWeights;
    suffix_name::String = "historical"
)

Compute ClimwipWeights based on RMSE distances between observational data and model-model pairs.

ModelWeights.Weights.climwipWeights — Method

climwipWeights(
    dists_indep_all::YAXArray, 
    dists_perform_all::YAXArray,
    config::ConfigWeights;
    suffix_name::Stirng = "climwip"
)

Compute weight for each model in multi-model ensemble according to approach from Brunner, Lukas, Angeline G. Pendergrass, Flavio Lehner, Anna L. Merrifield, Ruth Lorenz, and Reto Knutti. “Reduced Global Warming from CMIP6 Projections When Weighting Models by Performance and Independence.” Earth System Dynamics 11, no. 4 (November 13, 2020): 995–1012. https://doi.org/10.5194/esd-11-995-2020.

Arguments:

dists_indep_all::YAXArray: RMSEs between pairs of models for all

combinations of variables and diagnostics; has dimensions 'member1', 'member2', 'variable', 'diagnostic'.

dists_perform_all::YAXArray: RMSEs between model and observational data for all

combinations of variables and diagnostics; has dimensions 'member', 'variable', 'diagnostic'.

config::ConfigWeights: Parameters specifiying the relative contributions of each

combination of variable and diagnostic.

suffix_name::String: added to name of each type of weights, s.t. names are: wP-suffix,

wI-suffix, combined-suffix (default 'climwip').

ModelWeights.Weights.crps — Method

crps(samples::AbstractVector{<:Union{Missing, Number}}, obs::Number)

Return Continuous Ranked Probability Score (CRPS) for predicted samples and observation obs.

Quantifies how good the forecast samples are a prediction for the given observation. Use closed Form definition of CRPS.

Arguments:

samples::AbstractVector{<:Union{Missing, Number}}: vector of predictions.

obs::Number:: single observation.

ModelWeights.Weights.crps — Method

crps(samples::YAXArray{<:Union{Missing, Number}}, obs::YAXArray)

Compute Continuous Ranked Probability Score (CRPS) for predicted samples and observation obs.

Arguments:

samples: must have at least one more dimension than :model or :member (e.g. model x lon x lat)

obs: must have the same dimensions as samples except for the model dimension (:model or :member)

which is only present in samples.

ModelWeights.Weights.crpss — Method

crpss(samplesforecast::YAXArray, samplesbaseline::YAXArray, observations::YAXArray})

Compute Continuous Ranked Probability Skill Score for forecast with respect to baseline.

For every timestep there are predicted samples of the forecast (samples_forecast) and of the baseline samples_baseline (respectively one per model in the ensemble) and for every timestep there is exactly one observation. The CRPSS is based on the average CRPS across timesteps.

Arguments:

samples_forecast::YAXArray: must have dimension :time and :model or :member

samples_baseline::YAXArray: must have dimension :time and :model or :member

observations::YAXArray: must have only dimension :time

ModelWeights.Weights.crpss_weighted — Method

crpss_weighted(samples::YAXArray, observed::Number, weights::YAXArray)

Compute Continuous Ranked Probability Skill Score for weighted vs. unweighted predictions.

samples is one prediction per model in the ensemble, observed is the observed value.

Arguments:

samples::YAXArray: must have dimension :model or :member

observed::Number:

weights::YAXArray: must have same model dimension as samples (:model or :member)

n::Int=1000: length sample vector for computing weighted samples

ModelWeights.Weights.crpss_weighted — Method

crpss_weighted(samples::YAXArray, observations::YAXArray, weights::YAXArray; n::Int=1000)

Compute Continuous Ranked Probability Skill Score for weighted vs. unweighted predictions for timeseries.

For every timestep there are predicted samples, one per model in the ensemble and for every timestep there is exactly one observation. The CRPSS is based on the average CRPS across timesteps.

Arguments:

samples::YAXArray: must have dimension :time and :model or :member

observations::YAXArray: must have only dimension :time

weights::YAXArray: must have same model dimension as samples (:model or :member)

n::Int=1000: length sample vector for computing weighted samples

ModelWeights.Weights.distributeWeightsAcrossMembers — Method

distributeWeightsAcrossMembers(weights::YAXArray, members::Vector{String})

Equally distribute weight for each model over its members.

ModelWeights.Weights.equalWeights — Method

equalWeights(data::YAXArray; use_members::Bool=true)

Create a weight vector, with equal weight for each MODEL. Distribute weight across model members if dimension=:member and usemembers is true. If usemember is false, each model member is considered as standalone model and all receive the same weight.

ModelWeights.Weights.independenceWeights — Method

independenceWeights(Sij::YAXArray, sigmaS::Number)

Compute the independence weight for each model i with generalized distances between pairs of models, Sij.

\[w^{I}_i = a_0 \cdot \frac{1}{1 + \sum_{j \ne i} e^{-\left( \frac{S_{ij}}{\sigma_S} \right)^2}}\]

Arguments:

Sij: generalized distances for each pair of models with dimensions 'model1', 'model2'

sigmaS: free shape parameter for impact of independence weights

ModelWeights.Weights.performanceWeights — Method

performanceWeights(Di::AbstractArray, sigmaD::Number)

Compute the ClimWIP performance weight for each model i with generalized distances Di.

\[w^{P}_i = a_0 \cdot \frac{e^{-(\frac{D_i}{\sigma_D})^2}}{N}\]

Arguments:

Di::YAXArray: generalized distances Di for each model

sigmaD::Number: free shape parameter for impact of performance weights

ModelWeights.Weights.rps — Method

rps(p_hat::Vector{<:Number}, cat::Int)

ModelWeights.Weights.saveWeightsAsNCFile — Method

saveWeightsAsNCFile(weights::ClimWIP; target_path::String)

TODO: needs to be updated, deprecated

Arguments:

weights: ClimWIP object to be saved.

target_path: Path to where weights shall be stored.

ModelWeights.Weights.weightSamples — Method

weightSamples(samples::YAXArray, weights::YAXArray; n::Int = 1000)

Repeat each entry in samples proportional to respective value given in weights so that the returned vector has n entries.

Arguments:

samples::YAXArray: must have dimension :model or :member.

weights::YAXArray: must have same model dimension as samples (:model or :member).

n::Int = 1000: length of returned vector.

ModelWeights.Weights.weightedAvg — Method

weightedAvg(
    data::YAXArray; 
    weights::Union{YAXArray, Nothing} = nothing, 
    use_members_equal_weights::Bool = true
)

Compute the average values for each (lon,lat) grid point in data, weighted by weights.

If no weight vector is provided, unweighted average is computed.

Arguments:

data::YAXArray: must have dimension 'model' or 'member'

weights::Union{YAXArray, Nothing} = nothing: weights for models or individual members

use_members_equal_weights::Bool: if weights is nothing, all models receive

equal weight. Then, if use_members_equal_weights is true, the number of members per model is taken into account, e.g. ["m1#run1", "m1#run2", "m1#run3", "m2#run1"] yields [1/4, 1/4, 1/4, 1/4] if use_members_equal_weights = false and if use_members_equal_weights = true, [1/2 * 1/3, 1/2 * 1/3, 1/2 * 3, 1/2] = [1/6, 1/6, 1/6, 1/2].

ModelWeights.Weights.MMEWeights — Method

MMEWeights(weights::YAXArray, ids::Vector{String}, name::String)

Combine weights given in weights at ids into a normalized MMEWeight.

Arguments:

weights::YAXArray: Must have dimensions :weight and :model.

ids::Vector{String}: weights in weights where dimension :weights equals these ids will be combined.

name::String: Name of the new MMEWeights-instance.